% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regout.R
\name{add_base_vars}
\alias{add_base_vars}
\title{add_base_vars}
\usage{
add_base_vars(ht, no_binary_heading = TRUE)
}
\arguments{
\item{ht}{A huxtable.}

\item{no_binary_heading}{Do not add heading row for factors with 2 values.}
}
\description{
Add base (omitted) variables as a heading row to the regression.
}
