% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{add_latex_end}
\alias{add_latex_end}
\title{add_latex_end}
\usage{
add_latex_end(output)
}
\arguments{
\item{output}{The vector of LaTeX lines to which to add the closing lines to.
It is added to the beginning of the vector.}
}
\description{
"Close" the LaTeX document.
}
