% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{add_latex_preamble}
\alias{add_latex_preamble}
\title{add_latex_preamble}
\usage{
add_latex_preamble(output)
}
\arguments{
\item{output}{The vector of LaTeX lines to which to add the preamble to. It
is added to the beginning of the vector.}
}
\description{
Add a LaTeX preamble.
}
