% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regout.R
\name{add_reg_labels}
\alias{add_reg_labels}
\title{add_reg_labels}
\usage{
add_reg_labels(ht, var_list)
}
\arguments{
\item{ht}{A huxtable.}

\item{var_list}{A named list, where names are the original raw column names
and the values are the intended labels.}
}
\description{
Add regression labels.
}
