% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R, R/report_parameters.R,
%   R/report_statistics.R, R/report_table.R, R/report_text.R
\name{as.report}
\alias{as.report}
\alias{is.report}
\alias{as.report_parameters}
\alias{as.report_statistics}
\alias{as.report_table}
\alias{as.report_text}
\title{Create or test objects of class \link{report}.}
\usage{
as.report(text, table = NULL, plot = NULL, ...)

is.report(x)

as.report_parameters(x, summary = NULL, prefix = "  - ", ...)

as.report_statistics(x, summary = NULL, prefix = "  - ", ...)

as.report_table(x, ...)

as.report_text(x, ...)
}
\arguments{
\item{text}{Text obtained via \code{report_text()}}

\item{table}{Table obtained via \code{report_table()}}

\item{plot}{Plot obtained via \code{report_plot()}. Not yet implemented.}

\item{...}{Args to be saved as attributes.}

\item{x}{An arbitrary R object.}

\item{summary}{Add a summary as attribute (to be extracted via \code{summary()}).}

\item{prefix}{The prefix to be displayed in front of each parameter.}
}
\value{
A report object or a \code{TRUE/FALSE} value.
}
\description{
Allows to create or test whether an object is of the \code{report} class.
}
