% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_display_colours.R
\name{.colour_column_if}
\alias{.colour_column_if}
\title{Colour a column (Exported for test purposes)}
\usage{
.colour_column_if(x, name, condition = `<`, threshold = 0,
  colour_if = "green", colour_else = "red")
}
\description{
Colour a column (Exported for test purposes)
}
\keyword{internal}
