% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_display_colours.R
\name{.colour_columns}
\alias{.colour_columns}
\title{Colour a column (Exported for test purposes)}
\usage{
.colour_columns(x, names, colour = "red")
}
\description{
Colour a column (Exported for test purposes)
}
\keyword{internal}
