% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{.display.data.frame}
\alias{.display.data.frame}
\title{Dataframe Pretty Printing}
\usage{
\method{.display}{data.frame}(x, sep = " | ")
}
\description{
Dataframe Pretty Printing
}
\examples{
\dontrun{
x <- model_parameters(lm(Sepal.Length ~ Species * Sepal.Width, data=iris))
x <- .colour_column_if(x, name="CI_low", condition = `<`, threshold = 0,
                       colour_if = "red", colour_else = "green")
x <- .colour_column_if(x, name="p", condition = `<`, threshold = 0.05,
                       colour_if = "yellow", colour_else = NULL)
.display.data.frame(x)
}
}
\keyword{internal}
