% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_algorithm.R
\name{format_algorithm}
\alias{format_algorithm}
\title{Model Algorithm Formatting}
\usage{
format_algorithm(model)
}
\arguments{
\item{model}{A statistical model.}
}
\description{
Model Algorithm Formatting
}
\examples{
model <- lm(Sepal.Length ~ Species, data = iris)
format_algorithm(model)

if (require("lme4")) {
model <- lme4::lmer(Sepal.Length ~ Sepal.Width + (1|Species), data = iris)
format_algorithm(model)
}
}
