% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_bf.R
\name{format_bf}
\alias{format_bf}
\title{Bayes Factor Formatting}
\usage{
format_bf(bf, digits = 2)
}
\arguments{
\item{bf}{Bayes Factor.}

\item{digits}{Number of significant digits.}
}
\description{
Bayes Factor Formatting
}
\examples{
format_bf(1.20)
format_bf(c(1.20, 1557, 1, 8888))
}
