% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_CI.R
\name{format_CI}
\alias{format_CI}
\title{Confidence/Credible Interval (CI) Formatting}
\usage{
format_CI(CI_low, CI_high, CI = 95, digits = 2)
}
\arguments{
\item{CI_low}{Lower CI bound.}

\item{CI_high}{Upper CI bound.}

\item{CI}{CI level in percentage.}

\item{digits}{Number of significant digits.}
}
\description{
Confidence/Credible Interval (CI) Formatting
}
\examples{
format_CI(1.20, 3.57, CI = 90)
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
