% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_p.R
\name{format_p}
\alias{format_p}
\title{p-values Formatting}
\usage{
format_p(p, stars = FALSE, stars_only = FALSE)
}
\arguments{
\item{p}{value or vector of p-values.}

\item{stars}{Add significance stars.}

\item{stars_only}{Return only significance stars.}
}
\description{
p-values Formatting
}
\examples{
format_p(.02)
}
