% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_pd.R
\name{format_pd}
\alias{format_pd}
\title{Probability of direction (pd) Formatting}
\usage{
format_pd(pd, digits = 2)
}
\arguments{
\item{pd}{Lower CI bound.}

\item{digits}{Number of significant digits.}
}
\description{
Probability of direction (pd) Formatting
}
\examples{
format_pd(1.20)
}
