% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_rope.R
\name{format_rope}
\alias{format_rope}
\title{Confidence/Credible Interval (CI) Formatting}
\usage{
format_rope(rope_percentage, digits = 2)
}
\arguments{
\item{rope_percentage}{Value or vector of ROPE-percentages.}

\item{digits}{Number of significant digits.}
}
\description{
Confidence/Credible Interval (CI) Formatting
}
\examples{
format_rope(1.20, 3.57)
}
