% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_text.R, R/utils_text.R
\name{format_text}
\alias{format_text}
\alias{text_fullstop}
\alias{text_concatenate}
\alias{text_remove}
\alias{text_wrap}
\title{Convenient text formatting functionalities}
\usage{
format_text(text, sep = ", ", last = " and ", width = NULL, ...)

text_fullstop(text)

text_concatenate(text, sep = ", ", last = " and ")

text_remove(text, pattern = "", ...)

text_wrap(text, width = NULL, ...)
}
\arguments{
\item{text}{A character string.}

\item{sep}{Separator.}

\item{last}{Last separator.}

\item{width}{Positive integer giving the target column width for wrapping
lines in the output. Can be "auto", in which case it will select 90\% of the
default width.}

\item{...}{Other arguments to be passed to or from other functions.}

\item{pattern}{Character strings.}
}
\value{
A character string.
}
\description{
Convenience functions to manipulate and format text.
}
\examples{
library(report)

# Add full stop if missing
text_fullstop(c("something", "something else."))

# Smart concatenation
text_concatenate(c("First", "Second", "Last"))

# Remove parts of string
text_remove(c("one!", "two", "three!"), "!")

# Wrap text
long_text <- paste(rep("abc ", 100), collapse = "")
cat(text_wrap(long_text, width = 50))
}
