% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_text.R
\name{format_text_collapse}
\alias{format_text_collapse}
\title{Collapse Character Vector}
\usage{
format_text_collapse(x, sep = ", ", last = " and ")
}
\arguments{
\item{x}{Character vector.}

\item{sep}{Separator.}

\item{last}{Last separator.}
}
\description{
Collapse Character Vector
}
\examples{
format_text_collapse(c("A", "B", "C"))
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
