% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_value.R
\name{format_value}
\alias{format_value}
\title{Value Formatting}
\usage{
format_value(x, digits = 2)
}
\arguments{
\item{x}{Numeric value.}

\item{digits}{Number of significant digits.}
}
\description{
Value Formatting
}
\examples{
format_value(1.20)
format_value(1.2)
format_value(1.2012313)
format_value(c(0.0045, 234, -23))
}
