% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_d.R
\name{interpret_d}
\alias{interpret_d}
\title{Standardized Difference (Cohen's d) Interpretation}
\usage{
interpret_d(d, rules = "gignac2016")
}
\arguments{
\item{d}{Value or vector of d values.}

\item{rules}{Can be "gignac2016" (default), "cohen1988", "sawilowsky2009" or custom set of rules.}
}
\description{
Interpretation of indices using different sets of rules of thumb.
\href{https://easystats.github.io/report/articles/interpret_metrics.html#standardized-difference-d-cohens-d}{Click here} for details.
}
\examples{
interpret_d(.02)
interpret_d(c(.5, .02))
}
\references{
\itemize{
  \item Gignac, G. E., & Szodorai, E. T. (2016). Effect size guidelines for individual differences researchers. Personality and individual differences, 102, 74-78.
  \item Cohen, J. (1988). Statistical power analysis for the behavioural sciences.
  \item Sawilowsky, S. S. (2009). New effect size rules of thumb.
}
}
