% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_effective_sample.R
\name{interpret_effective_sample}
\alias{interpret_effective_sample}
\title{Effective Sample Size (ESS) Interpretation}
\usage{
interpret_effective_sample(effective_sample, rules = "burkner2017")
}
\arguments{
\item{effective_sample}{Value or vector of effective sample (ESS) values.}

\item{rules}{Can be "burkner2017" (default) or custom set of \code{\link{rules}}.}
}
\description{
Effective Sample Size (ESS) Interpretation
}
\examples{
interpret_effective_sample(1001)
interpret_effective_sample(c(852, 1200))
}
\references{
\itemize{
  \item Bürkner, P. C. (2017). brms: An R package for Bayesian multilevel models using Stan. Journal of Statistical Software, 80(1), 1-28.
}
}
