% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_omega_squared.R
\name{interpret_omega_squared}
\alias{interpret_omega_squared}
\title{Omega Squared Interpretation}
\usage{
interpret_omega_squared(omega_squared, rules = "field2013")
}
\arguments{
\item{omega_squared}{Value or vector of omega squared values.}

\item{rules}{Can be "field2013" (default) or custom set of rules.}
}
\description{
Omega Squared Interpretation
}
\examples{
interpret_omega_squared(.02)
interpret_omega_squared(c(.5, .02))
}
\references{
\itemize{
  \item{Field, A (2013) Discovering statistics using IBM SPSS Statistics. Fourth Edition. Sage:London.}
}
}
\seealso{
http://imaging.mrc-cbu.cam.ac.uk/statswiki/FAQ/effectSize
}
