% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_r.R
\name{interpret_r}
\alias{interpret_r}
\title{Correlation Coefficient Interpretation}
\usage{
interpret_r(r, rules = "cohen1988", direction = TRUE)
}
\arguments{
\item{r}{Value or vector of correlation coefficient.}

\item{rules}{Can be "cohen1988" (default), "evans1996" or custom set of rules.}

\item{direction}{Return the parameter's direction.}
}
\description{
Correlation Coefficient Interpretation
}
\examples{
interpret_r(r = .015)
interpret_r(r = c(.5, -.02))
#
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
