% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_rhat.R
\name{interpret_rhat}
\alias{interpret_rhat}
\title{Rhat Interpretation}
\usage{
interpret_rhat(rhat, rules = "vehtari2019")
}
\arguments{
\item{rhat}{Value or vector of Rhat values.}

\item{rules}{Can be "vehtari2019" (default), "gelman1992" or custom set of rules.}
}
\description{
Rhat Interpretation
}
\examples{
interpret_rhat(1.00)
interpret_rhat(c(1.5, 0.9))
}
\references{
\itemize{
  \item Gelman, A., & Rubin, D. B. (1992). Inference from iterative simulation using multiple sequences. Statistical science, 7(4), 457-472.
  \item Vehtari, A., Gelman, A., Simpson, D., Carpenter, B., & Bürkner, P. C. (2019). Rank-normalization, folding, and localization: An improved Rhat for assessing convergence of MCMC. arXiv preprint arXiv:1903.08008.
}
}
