% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_rope.R
\name{interpret_rope}
\alias{interpret_rope}
\title{ROPE percentage Interpretation}
\usage{
interpret_rope(rope_percentage, ci = 0.9, rules = "default")
}
\arguments{
\item{rope_percentage}{Value or vector of ROPE-percentages.}

\item{ci}{The Credible Interval (CI) probability, corresponding to the proportion of HDI, that was used. Can be \code{1} in the case of "full ROPE".}

\item{rules}{Can be \href{https://easystats.github.io/bayestestR/articles/guidelines.html}{"default"} or custom set of rules.}
}
\description{
ROPE percentage Interpretation
}
\examples{
interpret_rope(0, ci = 0.9)
interpret_rope(c(0.005, 0.99), ci = 1)
}
\references{
\href{https://easystats.github.io/bayestestR/articles/guidelines.html}{BayestestR's reporting guidelines}
}
