% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.lm.R
\name{model_data.lm}
\alias{model_data.lm}
\title{Retrieve Data from Linear Models}
\usage{
\method{model_data}{lm}(model, ...)
}
\arguments{
\item{model}{Object of class \link{lm}.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Attempts at retrieving data from a linear model.
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
