% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_model_methods.R
\name{model_data}
\alias{model_data}
\title{Retrieve Data}
\usage{
model_data(model, ...)
}
\arguments{
\item{model}{Statistical Model.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Attempts at retrieving data from model. See the documentation for your object's class:
\itemize{
 \item{\link[=model_data.lm]{lm}}
 }
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
