% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_model_methods.R
\name{model_description}
\alias{model_description}
\title{Model Description}
\usage{
model_description(model, ...)
}
\arguments{
\item{model}{Statistical Model.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Model description. See the documentation for your object's class:
\itemize{
 \item{\link[=model_description.lm]{lm}}
 }
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
