% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.lm.R
\name{model_initial.lm}
\alias{model_initial.lm}
\title{Linear Models Initial Parameters}
\usage{
\method{model_initial}{lm}(model, ...)
}
\arguments{
\item{model}{Object of class \link{lm}.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Initial parameters for linear models.
}
\examples{
model <- lm(Sepal.Length ~ Petal.Length * Species, data = iris)
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
