% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_model_methods.R
\name{model_initial}
\alias{model_initial}
\title{Initial Model Parameters}
\usage{
model_initial(model, ...)
}
\arguments{
\item{model}{Statistical Model.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Initial model parameters. See the documentation for your object's class:
\itemize{
 \item{\link[=model_initial.lm]{lm}}
 }
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
