% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.lm.R
\name{model_parameters.lm}
\alias{model_parameters.lm}
\title{Linear Models Description}
\usage{
\method{model_parameters}{lm}(model, ...)
}
\arguments{
\item{model}{Object of class \link{lm}.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Description for linear models.
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
