% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_model_methods.R
\name{model_parameters}
\alias{model_parameters}
\title{Model Parameters}
\usage{
model_parameters(model, ...)
}
\arguments{
\item{model}{Statistical Model.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Model parameters. See the documentation for your object's class:
\itemize{
 \item{\link[=model_parameters.lm]{lm}}
 }
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
