% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_model_methods.R
\name{model_performance}
\alias{model_performance}
\title{Compute Indices of Model Performance}
\usage{
model_performance(model, ...)
}
\arguments{
\item{model}{Statistical Model.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Indices of model performance. See the documentation for your object's class:
\itemize{
 \item{\link[=model_performance.lm]{lm}}
 }
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
