% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.lm.R
\name{model_table.lm}
\alias{model_table.lm}
\title{Create Tables for Linear Models}
\usage{
\method{model_table}{lm}(model, performance = TRUE, ...)
}
\arguments{
\item{model}{Object of class \link{lm}.}

\item{performance}{Add performance metrics on table.}

\item{...}{Arguments passed to or from other methods (see \link{model_parameters} and \link{model_performance}).}
}
\description{
Create Tables for Linear Models
}
\examples{
model <- lm(Sepal.Length ~ Petal.Length * Species, data = iris)
model_table(model)$table
}
