% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_table_bayesian.R
\name{model_table_bayesian}
\alias{model_table_bayesian}
\title{Create Tables for Bayesian Linear Models}
\usage{
model_table_bayesian(model, parameters, performance,
  performance_in_table = TRUE, ...)
}
\arguments{
\item{model}{Bayesian models.}

\item{parameters}{Parameters table \link{model_parameters}.}

\item{performance}{Performance table \link{model_performance}.}

\item{performance_in_table}{Add performance metrics on table.}

\item{...}{Arguments passed to or from other methods (see \link{model_parameters} and \link{model_performance}).}
}
\description{
Create Tables for Bayesian Linear Models
}
\keyword{internal}
