% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.glm.R
\name{model_values.glm}
\alias{model_values.glm}
\title{General Linear Models Values}
\usage{
\method{model_values}{glm}(model, ci = 0.95, standardize = TRUE,
  effsize = "chen2010", performance_in_table = TRUE,
  performance_metrics = "all", bootstrap = FALSE, ...)
}
\arguments{
\item{model}{Object of class \link{lm}.}

\item{ci}{Confidence Interval (CI) level. Default to 0.95 (95\%).}

\item{standardize}{Standardized coefficients. See \code{\link[parameters:model_parameters.lm]{model_parameters}}.}

\item{effsize}{\href{https://easystats.github.io/report/articles/interpret_metrics.html}{Interpret the standardized parameters} using a set of rules. Can be "cohen1988" (default for linear models), "chen2010" (default for logistic models), "sawilowsky2009", NULL, or a custom set of \link{rules}.}

\item{performance_in_table}{Add performance metrics in table.}

\item{performance_metrics}{See \code{\link[performance:model_performance.lm]{model_performance}}.}

\item{bootstrap}{See \code{\link[parameters:model_parameters.lm]{model_parameters}}.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Extract all values of general linear models.
}
