% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{print.report}
\alias{print.report}
\title{Print the results.}
\usage{
\method{print}{report}(x, full = FALSE, ...)
}
\arguments{
\item{x}{An object of class \link{report}}

\item{full}{Show the full report.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print the results.
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
