% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{print.report_table}
\alias{print.report_table}
\title{Parameters table printing}
\usage{
\method{print}{report_table}(x, digits = 2, ...)
}
\arguments{
\item{x}{Object of class \code{table_report}.}

\item{digits}{Number of digits.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Parameters table printing
}
