% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_performance_r2_MLM.R
\name{r2_MLM}
\alias{r2_MLM}
\title{R-squared for Multilevel Models (r2MLM; Rights & Sterba, 2018)}
\usage{
r2_MLM(data, Tau, sigma2, within_covs = NULL, between_covs = NULL,
  random_covs = NULL, gamma_w = NULL, gamma_b = NULL,
  has_intercept = TRUE, clustermeancentered = TRUE)
}
\arguments{
\item{data}{dataset with rows denoting observations and columns denoting variables}

\item{Tau}{random effect covariance matrix; note that the first row/column denotes the intercept variance and covariances (if intercept is fixed, set all to 0) and each subsequent row/column denotes a given random slope’s variance and covariances (to be entered in the order listed by random_covs)}

\item{sigma2}{level-1 residual variance}

\item{within_covs}{list of numbers corresponding to the columns in the dataset of the level-1 predictors used in the MLM (if none used, set to NULL)}

\item{between_covs}{list of numbers corresponding to the columns in the dataset of the level-2 predictors used in the MLM (if none used, set to NULL)}

\item{random_covs}{list of numbers corresponding to the columns in the dataset of the level-1 predictors that have random slopes in the MLM (if no random slopes, set to NULL)}

\item{gamma_w}{vector of fixed slope estimates for all level-1 predictors, to be entered in the order of the predictors listed by within_covs (if none, set to NULL)}

\item{gamma_b}{vector of fixed intercept estimate (if applicable; see has_intercept below) and fixed slope estimates for all level-2 predictors, to be entered intercept first (if applicable) followed by level-2 slopes in the order listed by between_covs (if none, set to NULL)}

\item{has_intercept}{if set to TRUE, the first element of gamma_b is assumed to be the fixed intercept estimate; if set to FALSE, the first element of gamma_b is assumed to be the first fixed level-2 predictor slope; set to TRUE by default}

\item{clustermeancentered}{if set to TRUE, all level-1 predictors (indicated by the within_covs list) are assumed to be cluster-mean-centered and function will output all decompositions; if set to FALSE, function will output only total decompositions (see Description above); set to TRUE by default}
}
\description{
This function reads in raw data and multilevel model (MLM) parameter estimates and outputs all relevant  measures and barchart decompositions. That is, when predictors are cluster-mean-centered, all R-squareds in Table 1 and decompositions in Figure 1 are outputted (see Rights & Sterba, 2018). When predictors are not cluster-mean-centered, the total R-squareds from Table 5,  as well as barchart decompositions are outputted. Any number of level-1 and/or level-2 predictors is supported. Any of the level-1 predictors can have random slopes.
}
\examples{
print("no examples yet")
}
\references{
- Rights & Sterba (2018). Quantifying explained variance in multilevel models: An integrative framework for defining R-squared measures. Psychological Methods.
- Rights & Cole (2018) Effect Size Measures for Multilevel Models in Clinical Child and Adolescent Research: New R-Squared Methods and Recommendations, Journal of Clinical Child & Adolescent Psychology, 47:6.
}
\seealso{
- https://my.vanderbilt.edu/jasonrights/software/r2mlm/
- https://mgb-research.netlify.com/post/visualizing-variance-in-multilevel-models-using-the-riverplot-package/
}
\author{
Jason D. Right, \href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
