% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regout.R
\name{regout}
\alias{regout}
\title{regout}
\usage{
regout(..., heading_rows = 1)
}
\arguments{
\item{...}{arguments to be passed to huxreg.}

\item{heading_rows}{The number of rows which are headings.}
}
\description{
A wrapper around huxreg which allows specification of the number of initial
heading rows. It also adds model term labels and xlevels as attributes.
}
