% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.array.R
\name{report.array}
\alias{report.array}
\title{Reporting Arrays}
\usage{
\method{report}{array}(
  x,
  n = FALSE,
  centrality = "mean",
  dispersion = TRUE,
  range = TRUE,
  distribution = FALSE,
  levels_percentage = "auto",
  digits = 2,
  n_entries = 3,
  missing_percentage = "auto",
  ...
)
}
\arguments{
\item{x}{The R object that you want to report (see list of of supported
objects above).}

\item{n}{Include number of observations for each individual variable.}

\item{centrality}{Character vector, indicating the index of centrality
(either \code{"mean"} or \code{"median"}).}

\item{dispersion}{Show index of dispersion (\link{sd} if \code{centrality = "mean"}, or
\link{mad} if \code{centrality = "median"}).}

\item{range}{Show range.}

\item{distribution}{Show kurtosis and skewness.}

\item{levels_percentage}{Show characters entries and factor levels by number
or percentage. If "auto", then will be set to number and percentage if the
length if n observations larger than 100.}

\item{digits}{Number of significant digits.}

\item{n_entries}{Number of different character entries to show. Can be "all".}

\item{missing_percentage}{Show missing by number (default) or percentage. If
"auto", then will be set to number and percentage if the length if n
observations larger than 100.}

\item{...}{Arguments passed to or from other methods.}
}
\value{
An object of class \code{\link[=report]{report()}}.
}
\description{
Create reports for arrays.
}
\examples{
library(report)

r <- report(iris3)
r
summary(r)
as.data.frame(r)
summary(as.data.frame(r))
}
