% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.bayestestR.R
\name{report.bayesfactor_inclusion}
\alias{report.bayesfactor_inclusion}
\title{Inclusion Bayes factor Report}
\usage{
\method{report}{bayesfactor_inclusion}(model, rules = "jeffreys1961",
  ...)
}
\arguments{
\item{model}{Object of class \code{bayesfactor_inclusion}.}

\item{rules}{Can be "jeffreys1961" (default), "raftery1995" or custom set of \code{\link{rules}}.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Create a report of inclusion Bayes factors.
}
\examples{
\donttest{
library(bayestestR)
library(report)

mo0 <- lm(Sepal.Length ~ 1, data = iris)
mo1 <- lm(Sepal.Length ~ Species, data = iris)
mo2 <- lm(Sepal.Length ~ Species + Petal.Length, data = iris)
mo3 <- lm(Sepal.Length ~ Species * Petal.Length, data = iris)

BFmodels <- bayesfactor_models(mo1, mo2, mo3, denominator = mo0)
inc_bf <- bayesfactor_inclusion(BFmodels, prior_odds = c(1,2,3), match_models = TRUE)

bf_report <- report(inc_bf, digits = 3)
to_table(bf_report)
to_text(bf_report, full = FALSE)

}
}
\seealso{
report
}
