% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.brmsfit.R
\name{report.brmsfit}
\alias{report.brmsfit}
\alias{report_effectsize.brmsfit}
\title{Reporting Bayesian Models from brms}
\usage{
\method{report}{brmsfit}(x, ...)

\method{report_effectsize}{brmsfit}(x, effectsize_method = "basic", ...)
}
\arguments{
\item{x}{Object of class \code{lm} or \code{glm}.}

\item{...}{Arguments passed to or from other methods.}

\item{effectsize_method}{Method for computing effect sizes. For \code{brmsfit} objects,
defaults to \code{"basic"} (faster, no refitting) instead of \code{"refit"} to improve
performance with large Bayesian models. See documentation for
\code{\link[effectsize:effectsize]{effectsize::effectsize()}}.}
}
\value{
An object of class \code{\link[=report]{report()}}.
}
\description{
Create reports for Bayesian models. The description of the parameters
follows the Sequential Effect eXistence and sIgnificance Testing framework
(see \link[bayestestR:sexit]{SEXIT documentation}).
}
\details{
Message from the \code{rstan} package: "To avoid recompilation of unchanged
Stan programs, we recommend calling \code{rstan_options(auto_write = TRUE)}"
}
\examples{
\dontshow{if (require("brms", quietly = TRUE)) withAutoprint(\{ # examplesIf}
\donttest{
# Bayesian models
library(brms)
model <- suppressWarnings(brm(mpg ~ qsec + wt,
  data = mtcars,
  refresh = 0, iter = 300
))
r <- report(model, verbose = FALSE)
r
summary(r)
as.data.frame(r)
summary(as.data.frame(r))
}
\dontshow{\}) # examplesIf}
}
\seealso{
Specific components of reports (especially for stats models):
\itemize{
\item \code{\link[=report_table]{report_table()}}
\item \code{\link[=report_parameters]{report_parameters()}}
\item \code{\link[=report_statistics]{report_statistics()}}
\item \code{\link[=report_effectsize]{report_effectsize()}}
\item \code{\link[=report_model]{report_model()}}
\item \code{\link[=report_priors]{report_priors()}}
\item \code{\link[=report_random]{report_random()}}
\item \code{\link[=report_performance]{report_performance()}}
\item \code{\link[=report_info]{report_info()}}
\item \code{\link[=report_text]{report_text()}}
}

Other types of reports:
\itemize{
\item \code{\link[=report_system]{report_system()}}
\item \code{\link[=report_packages]{report_packages()}}
\item \code{\link[=report_participants]{report_participants()}}
\item \code{\link[=report_sample]{report_sample()}}
\item \code{\link[=report_date]{report_date()}}
}

Methods:
\itemize{
\item \code{\link[=as.report]{as.report()}}
}

Template file for supporting new models:
\itemize{
\item \code{\link[=report.default]{report.default()}}
}
}
