% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.correlation.R
\name{report.correlation}
\alias{report.correlation}
\title{Correlation Report}
\usage{
\method{report}{correlation}(model, effsize = "cohen1988", ...)
}
\arguments{
\item{model}{Object of class correlation}

\item{effsize}{Effect size interpretation set of rules.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Create a report of a correlation object.
}
\examples{
model <- correlation(iris)
report(model)
}
\seealso{
report
}
