% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_regression.R
\name{report.default}
\alias{report.default}
\title{(General) Linear Models Report}
\usage{
\method{report}{default}(
  model,
  interpretation = "default",
  ci = 0.95,
  standardize = "refit",
  standardize_robust = FALSE,
  bootstrap = FALSE,
  iterations = 500,
  performance_metrics = "all",
  ...
)
}
\arguments{
\item{model}{Model object.}

\item{interpretation}{\href{https://easystats.github.io/effectsize/articles/interpret.html}{Interpret the standardized parameters} using a set of rules. Default corresponds to "cohen1988" for linear models and "chen2010" for logistic models.}

\item{ci}{Confidence Interval (CI) level. Default to 0.95 (95\%).}

\item{standardize}{The method used for standardizing the parameters. Can be \code{"refit"}, \code{"posthoc"}, \code{"smart"}, \code{"basic"} or \code{NULL} (default) for no standardization. See 'Details' in \code{\link[effectsize]{standardize_parameters}}. Note that robust estimation (i.e. \code{robust=TRUE}) of standardized parameters only works when \code{standardize="refit"}.}

\item{standardize_robust}{Logical, if \code{TRUE}, robust standard errors are calculated (if possible), and confidence intervals and p-values are based on these robust standard errors.}

\item{bootstrap}{Should estimates be based on bootstrapped model? If \code{TRUE}, then arguments of \link[parameters:model_parameters.stanreg]{Bayesian regressions} apply (see also \code{\link[parameters:bootstrap_parameters]{bootstrap_parameters()}}).}

\item{iterations}{The number of bootstrap replicates. This only apply in the case of bootstrapped frequentist models.}

\item{performance_metrics}{See \code{\link[performance:model_performance.lm]{model_performance}}.}

\item{...}{Arguments passed to or from other methods. For instance, when \code{bootstrap = TRUE}, arguments like \code{ci_method} are passed down to \code{\link[bayestestR]{describe_posterior}}.}
}
\value{
A list-object of class \code{report}, which contains further list-objects
with a short and long description of the model summary, as well as a short
and long table of parameters and fit indices.
}
\description{
Create a report of a (general) linear model.
}
\examples{
library(report)

model <- lm(Sepal.Length ~ Petal.Length * Species, data = iris)
r <- report(model)
text_short(r)
text_long(r)
table_short(r)
table_long(r)


model <- glm(vs ~ disp, data = mtcars, family = "binomial")
r <- report(model)
text_short(r)
text_long(r)
table_short(r)
table_long(r)
}
\seealso{
\code{\link{table_short}} or \code{\link{text_short}} to access the
related content of the report-object.
}
