% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.estimate_contrasts.R
\name{report.estimate_contrasts}
\alias{report.estimate_contrasts}
\alias{report_table.estimate_contrasts}
\alias{report_text.estimate_contrasts}
\title{Reporting \code{estimate_contrasts} objects}
\usage{
\method{report}{estimate_contrasts}(x, ...)

\method{report_table}{estimate_contrasts}(x, ...)

\method{report_text}{estimate_contrasts}(x, table = NULL, ...)
}
\arguments{
\item{x}{Object of class \code{estimate_contrasts}.}

\item{...}{Arguments passed to or from other methods.}

\item{table}{Provide the output of  \code{report_table()} to avoid its
re-computation.}
}
\value{
An object of class \code{\link[=report]{report()}}.
}
\description{
Create reports for \code{estimate_contrasts} objects.
}
\examples{
\dontshow{if (all(insight::check_if_installed(c("modelbased", "marginaleffects", "collapse", "Formula"), quietly = TRUE))) withAutoprint(\{ # examplesIf}
library(modelbased)
model <- lm(Sepal.Width ~ Species, data = iris)
contr <- estimate_contrasts(model)
report(contr)
\dontshow{\}) # examplesIf}
}
\seealso{
Specific components of reports (especially for stats models):
\itemize{
\item \code{\link[=report_table]{report_table()}}
\item \code{\link[=report_parameters]{report_parameters()}}
\item \code{\link[=report_statistics]{report_statistics()}}
\item \code{\link[=report_effectsize]{report_effectsize()}}
\item \code{\link[=report_model]{report_model()}}
\item \code{\link[=report_priors]{report_priors()}}
\item \code{\link[=report_random]{report_random()}}
\item \code{\link[=report_performance]{report_performance()}}
\item \code{\link[=report_info]{report_info()}}
\item \code{\link[=report_text]{report_text()}}
}

Other types of reports:
\itemize{
\item \code{\link[=report_system]{report_system()}}
\item \code{\link[=report_packages]{report_packages()}}
\item \code{\link[=report_participants]{report_participants()}}
\item \code{\link[=report_sample]{report_sample()}}
\item \code{\link[=report_date]{report_date()}}
}

Methods:
\itemize{
\item \code{\link[=as.report]{as.report()}}
}

Template file for supporting new models:
\itemize{
\item \code{\link[=report.default]{report.default()}}
}
}
