% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.estimate.R
\name{report.estimateContrasts}
\alias{report.estimateContrasts}
\title{Contrasts Estimate Report}
\usage{
\method{report}{estimateContrasts}(model, ...)
}
\arguments{
\item{model}{Object of class \code{estimate}}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Create a report of an contrasts estimate object.
}
\examples{
\dontrun{
library(estimate)
library(rstanarm)

model <- stan_glm(Sepal.Width ~ Species * Petal.Width, data = iris)
report(estimate_contrasts(model))
report(estimate_contrasts(model, fixed = "Petal.Width"))
report(estimate_contrasts(model, modulate = "Petal.Width", length = 4))
}

}
\seealso{
report
}
