% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.factor.R
\name{report.factor}
\alias{report.factor}
\title{Factor (Categorical Vector) Report}
\usage{
\method{report}{factor}(x, levels_percentage = TRUE,
  missing_percentage = FALSE, ...)
}
\arguments{
\item{x}{a categorical vector.}

\item{levels_percentage}{show factor levels by percentage (default) or number.}

\item{missing_percentage}{show missings by number (default) or percentage}

\item{...}{arguments passed to or from other methods.}
}
\description{
Create a report of a categorical vector.
}
\examples{
x <- factor(rep(c("A", "B", "C"), 10))
report(x)
report(x, levels_percentage = FALSE, missing_percentage = TRUE)
to_fulltext(report(x))
to_table(report(x))
to_fulltable(report(x))
}
\seealso{
report
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
