% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.glm.R
\name{report.glm}
\alias{report.glm}
\title{Linear Models Report}
\usage{
\method{report}{glm}(model, ci = 0.95, standardize = TRUE,
  effsize = "chen2010", performance_in_table = TRUE,
  performance_metrics = "all", bootstrap = FALSE, ...)
}
\arguments{
\item{model}{Object of class \link{lm}.}

\item{ci}{Confidence Interval (CI) level. Default to 0.95 (95\%).}

\item{standardize}{Standardized coefficients. See \code{\link[parameters:model_parameters.lm]{model_parameters}}.}

\item{effsize}{Interpret the log odds ratio using a set of rules. Can be "chen2010" (default), "cohen1988", NULL, or a custom set of \link{rules}.}

\item{performance_in_table}{Add performance metrics on table.}

\item{performance_metrics}{See \code{\link[performance:model_performance.lm]{model_performance}}.}

\item{bootstrap}{See \code{\link[parameters:model_parameters.lm]{model_parameters}}.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Create a report of a linear model.
}
\examples{
model <- lm(Sepal.Length ~ Petal.Length * Species, data = iris)
r <- report(model)
to_text(r)
to_fulltext(r)
to_table(r)
to_fulltable(r)
}
