% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.merMod.R
\name{report.lmerMod}
\alias{report.lmerMod}
\alias{report.merMod}
\title{Mixed Models Report}
\usage{
\method{report}{lmerMod}(model, ci = 0.95, standardize = "refit",
  standardize_robust = FALSE, effsize = "cohen1988",
  performance_in_table = TRUE, performance_metrics = "all",
  bootstrap = FALSE, p_method = "wald", ci_method = "wald", ...)

\method{report}{merMod}(model, ci = 0.95, standardize = "refit",
  standardize_robust = FALSE, effsize = "chen2010",
  performance_in_table = TRUE, performance_metrics = "all",
  bootstrap = FALSE, ci_method = "wald", ...)
}
\arguments{
\item{model}{Mixed model.}

\item{ci}{Confidence Interval (CI) level. Default to 0.95 (95\%).}

\item{standardize}{Add standardized parameters. Can be FALSE or a character indicating the standardization method (see \code{\link{parameters_standardize}}).}

\item{standardize_robust}{Robust standardization. See \code{\link{parameters_standardize}}.}

\item{effsize}{\href{https://easystats.github.io/report/articles/interpret_metrics.html}{Interpret the standardized parameters} using a set of rules. Can be "cohen1988" (default for linear models), "chen2010" (default for logistic models), "sawilowsky2009", NULL, or a custom set of \link{rules}.}

\item{performance_metrics}{See \code{\link[performance:model_performance.lm]{model_performance}}.}

\item{bootstrap}{Should estimates be based on bootsrapped model? If TRUE, then arguments of \link[=model_parameters.stanreg]{Bayesian regressions} apply.}

\item{p_method}{Method for computing p values. See \link{p_value}.}

\item{ci_method}{Method for computing confidence intervals (CI). See \link{ci}.}

\item{...}{Arguments passed to or from other methods (e.g., to \link[=standardize.lm]{standardize}).}
}
\description{
Create a report of a mixed model.
}
\examples{
\dontrun{
model <- lmer(Sepal.Length ~ Petal.Length + (1 | Species), data = iris)
r <- report(model)
to_text(r)
to_fulltext(r)
to_table(r)
to_fulltable(r)
}
}
