% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_ai.R
\name{report_ai}
\alias{report_ai}
\alias{report_ai.merMod}
\alias{report_ai.glmmTMB}
\title{Generate AI-optimized reports}
\usage{
report_ai(x, ...)

\method{report_ai}{merMod}(x, ...)

\method{report_ai}{glmmTMB}(x, ...)
}
\arguments{
\item{x}{A statistical model.}

\item{...}{Arguments passed to other functions, like \code{parameters::model_parameters()},
\code{performance::model_performance()} or \code{insight::format_table()}.}
}
\value{
A character vector of class \code{report_ai} containing the formatted text.
}
\description{
This function is designed to produce AI-optimized output for statistical models.
It strikes a careful balance between comprehensiveness, specificity, and compactness.
The primary goal is to provide a Large Language Model (LLM) or AI agent with the
clearest and most relevant analytical information at the lowest possible token cost.
}
\examples{
m <- lm(mpg ~ wt + hp, data = mtcars)
report_ai(m)
\dontshow{if (requireNamespace("lme4", quietly = TRUE)) withAutoprint(\{ # examplesIf}
\donttest{
m <- lme4::lmer(Reaction ~ Days + (1 | Subject), data = lme4::sleepstudy)
report_ai(m)
}
\dontshow{\}) # examplesIf}
\dontshow{if (requireNamespace("glmmTMB", quietly = TRUE)) withAutoprint(\{ # examplesIf}
\donttest{
m <- glmmTMB::glmmTMB(count ~ mined + (1 | site), family = poisson(), data = glmmTMB::Salamanders)
report_ai(m)
}
\dontshow{\}) # examplesIf}
}
