% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_b.R
\name{report_b}
\alias{report_b}
\title{Generate an in-text report of regression coeffiecients}
\usage{
report_b(coefficient = NULL, markdown = T, standardize = F)
}
\arguments{
\item{coefficient}{object containing regression coefficient to report}

\item{markdown}{should the output be formatted for Markdown (e.g., *b* vs b)
or plain text [T/F]?}

\item{standardize}{is the coefficient undstandardized [F] or standardized [T]}
}
\value{
unstandardized or standardized regression coefficient
rounded to three decimal places (e.g., b = .123, beta = .123).
}
\description{
Generate an in-text report of regression coeffiecients
}
\examples{
x <- c(1:25)
y = c(1, 1, 2, 3, 4, 5, 6, 2, 4, 6, 7, 2, 7, 3, 1, 1, 1, 3, 4, 6, 7, 7, 1, 4, 1)
model <- lm(y ~ x)
coef <- summary(model)$coefficients[2, 1]
report_b(coef, standardize = FALSE, markdown = FALSE)
}
