% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-blr.R
\name{report_blr}
\alias{report_blr}
\title{Logistic regression report}
\usage{
report_blr(use_yaml = FALSE)
}
\arguments{
\item{use_yaml}{Use YAML file to specify report details.}
}
\description{
Generate detailed logistic regression report.
}
\examples{
\dontrun{
model_data <- blorr::bank_marketing

# interactive
report_blr()

# use yaml file
report_blr(use_yaml = TRUE)
}

}
