% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_misc.R
\name{report_date}
\alias{report_date}
\alias{report_story}
\title{Miscellaneous reports}
\usage{
report_date(...)

report_story(...)
}
\arguments{
\item{...}{Arguments passed to or from other methods.}
}
\value{
Objects of class \code{\link[=report_text]{report_text()}}.
}
\description{
Other convenient or totally useless reports.
}
\examples{
library(report)

report_date()
summary(report_date())
report_story()
}
\seealso{
Specific components of reports (especially for stats models):
\itemize{
\item \code{\link[=report_table]{report_table()}}
\item \code{\link[=report_parameters]{report_parameters()}}
\item \code{\link[=report_statistics]{report_statistics()}}
\item \code{\link[=report_effectsize]{report_effectsize()}}
\item \code{\link[=report_model]{report_model()}}
\item \code{\link[=report_priors]{report_priors()}}
\item \code{\link[=report_random]{report_random()}}
\item \code{\link[=report_performance]{report_performance()}}
\item \code{\link[=report_info]{report_info()}}
\item \code{\link[=report_text]{report_text()}}
}

Other types of reports:
\itemize{
\item \code{\link[=report_system]{report_system()}}
\item \code{\link[=report_packages]{report_packages()}}
\item \code{\link[=report_participants]{report_participants()}}
\item \code{\link[=report_sample]{report_sample()}}
\item \code{\link[=report_date]{report_date()}}
}

Methods:
\itemize{
\item \code{\link[=as.report]{as.report()}}
}

Template file for supporting new models:
\itemize{
\item \code{\link[=report.default]{report.default()}}
}
}
