% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_LRT.R
\name{report_LRT}
\alias{report_LRT}
\title{Generate an in-text report of a Likelihood Ratio Test}
\usage{
report_LRT(result = NULL, format = "rmarkdown", inline = T)
}
\arguments{
\item{result}{Results from likelihood ratio test to report}

\item{format}{How should the output be formatted? Available options are:
\code{"plain"}, \code{"latex"}, \code{"rmarkdown"}}

\item{inline}{Should non-plain results be formatted for inline (\code{TRUE})
or for copy-pasting (\code{FALSE})?
(default is \code{TRUE})}
}
\value{
Formatted results containing chi-square statistic, degrees of freedom,
and p-value.
}
\description{
Generate an in-text report of a Likelihood Ratio Test
}
\examples{
\dontrun{
politeness <- read.csv("http://www.bodowinter.com/tutorial/politeness_data.csv")
library(lme4)
max.model <- lmer(frequency ~ attitude + (1 | subject) + (1 | scenario), data=politeness)
reduced.model <- lmer(frequency ~ attitude + (1 | subject), data=politeness)
results <- anova(reduced.model, max.model, refit = F)[2, ]
report_LRT(result = results, format = 'rmarkdown')
}
}
