% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_msd.R
\name{report_msd}
\alias{report_msd}
\title{Generate an in-text report of means and standard deviations}
\usage{
report_msd(var = NULL, markdown = T)
}
\arguments{
\item{var}{variable to report descriptives for}

\item{markdown}{should the output be formatted for Markdown (e.g., *M* vs M)
or plain text [T/F]?}
}
\value{
Mean and SD reported to two decimal places
}
\description{
Generate an in-text report of means and standard deviations
}
\examples{
age <- rnorm(110, 21, 3.4)
report_msd(age)
}
