% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_p.R
\name{report_p}
\alias{report_p}
\title{Automate the reporting of p-values}
\usage{
report_p(pvalue = NULL, format = "rmarkdown", inline = T)
}
\arguments{
\item{pvalue}{object containing pvalue to report}

\item{format}{How should the output be formatted? Available options are:
\code{"plain"}, \code{"latex"}, \code{"rmarkdown"}}

\item{inline}{Should non-plain results be formatted for inline (\code{TRUE})
or for copy-pasting (\code{FALSE})?
(default is \code{TRUE})}
}
\value{
p-value rounded to three decimal places if greater than .001. If less than,
.001 defaults to p < .001.
}
\description{
Automate the reporting of p-values
}
\examples{
report_p(0.001567)
}
