% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_r.R
\name{report_r}
\alias{report_r}
\title{Generate an in-text report of correlation results}
\usage{
report_r(corcoef = NULL, markdown = T)
}
\arguments{
\item{corcoef}{object containing correlation coefficient}

\item{markdown}{should the output be formatted for Markdown (e.g., *r* vs r)
or plain text [T/F]?}
}
\value{
correlation coefficient reported to three decimal places (no leading zero)
}
\description{
Generate an in-text report of correlation results
}
\examples{
x1 <- rnorm(200, 10, 2)
x2 <- rnorm(200, 15, 1.5)
r_coef <- cor(x1, x2)
report_r(r_coef)
}
