% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-rfm.R
\name{report_rfm}
\alias{report_rfm}
\title{RFM report}
\usage{
report_rfm(use_yaml = FALSE)
}
\arguments{
\item{use_yaml}{Use YAML file to specify report details.}
}
\description{
Generate recency, frequency and monetary value analysis report.
}
\examples{
\dontrun{
rfm_data <- rfm::rfm_data_orders

# interactive
report_rfm()

# use yaml file
report_rfm(use_yaml = TRUE)

}

}
