% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_t.R
\name{report_t}
\alias{report_t}
\title{Generate an in-text report of t-test results}
\usage{
report_t(tresults = NULL, format = "rmarkdown", inline = T)
}
\arguments{
\item{tresults}{object containing t-test results from t.test()}

\item{format}{How should the output be formatted? Available options are:
\code{"plain"}, \code{"latex"}, \code{"rmarkdown"}}

\item{inline}{Should non-plain results be formatted for inline or for copy-pasting?}
}
\value{
t-value reported to two decimal places, degrees of freedom, and p-value reported
to three decimal places
}
\description{
Generate an in-text report of t-test results
}
\examples{
t1 <- rnorm(110, 50, 10)
t2 <- rnorm(110, 55, 12)
tresult <- t.test(t1, t2)
report_t(tresult)
}
