% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.check_outliers.R
\name{report_text.check_outliers}
\alias{report_text.check_outliers}
\alias{report.check_outliers}
\title{Reporting outlier detection results}
\usage{
\method{report_text}{check_outliers}(x, ...)

\method{report}{check_outliers}(x, ...)
}
\arguments{
\item{x}{Object of class \code{check_outliers} as returned by
\code{\link[performance:check_outliers]{performance::check_outliers()}}.}

\item{...}{Arguments passed to or from other methods.}
}
\value{
An object of class \code{\link[=report_text]{report_text()}}.
}
\description{
Create a brief textual report for outlier detection results from
\code{\link[performance:check_outliers]{performance::check_outliers()}}.
}
\examples{
\dontshow{if (requireNamespace("performance", quietly = TRUE)) withAutoprint(\{ # examplesIf}
\donttest{
library(report)
library(performance)

mt2 <- rbind(
  mtcars[, c("mpg", "disp", "hp")],
  data.frame(mpg = c(37, 40), disp = c(300, 400), hp = c(110, 120))
)
model <- lm(disp ~ mpg + hp, data = mt2)
rez <- performance::check_outliers(model)
report(rez)
}
\dontshow{\}) # examplesIf}
}
\seealso{
Specific components of reports (especially for stats models):
\itemize{
\item \code{\link[=report_table]{report_table()}}
\item \code{\link[=report_parameters]{report_parameters()}}
\item \code{\link[=report_statistics]{report_statistics()}}
\item \code{\link[=report_effectsize]{report_effectsize()}}
\item \code{\link[=report_model]{report_model()}}
\item \code{\link[=report_priors]{report_priors()}}
\item \code{\link[=report_random]{report_random()}}
\item \code{\link[=report_performance]{report_performance()}}
\item \code{\link[=report_info]{report_info()}}
\item \code{\link[=report_text]{report_text()}}
}

Other types of reports:
\itemize{
\item \code{\link[=report_system]{report_system()}}
\item \code{\link[=report_packages]{report_packages()}}
\item \code{\link[=report_participants]{report_participants()}}
\item \code{\link[=report_sample]{report_sample()}}
\item \code{\link[=report_date]{report_date()}}
}

Methods:
\itemize{
\item \code{\link[=as.report]{as.report()}}
}

Template file for supporting new models:
\itemize{
\item \code{\link[=report.default]{report.default()}}
}
}
