% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret.R
\name{rules}
\alias{rules}
\title{Interpretation Grid}
\usage{
rules(breakpoints, labels, if_lower = TRUE)
}
\arguments{
\item{breakpoints}{Vector of value break points (edges defining categories).}

\item{labels}{Labels associated with each category. Must contain one label more than breakpoints.}

\item{if_lower}{If true, each label will be given if the value is strictly lower than its breakpoint. The contrary if false.}
}
\description{
Create a container for interpretation rules of thumb. See \link{interpret}.
}
\examples{
rules(c(0.05), c("significant", "not significant"), if_lower = TRUE)
}
