% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_cite_packages.R
\name{show_packages}
\alias{show_packages}
\title{Cite Loaded Packages}
\usage{
show_packages(session)
}
\arguments{
\item{session}{A \link[=sessionInfo]{sessionInfo} object.}
}
\description{
Citation table of loaded packages (show_packages includes version and name, and cite_packages includes only the citation).
}
\examples{
\dontrun{
show_packages(sessionInfo())
cite_packages(sessionInfo())
}

}
\author{
\href{https://github.com/DominiqueMakowski}{Dominique Makowski}
}
