% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize.factor}
\alias{standardize.factor}
\title{Standardization}
\usage{
\method{standardize}{factor}(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Standardize (scale and reduce, Z-score) X so that the values are expressed in terms of standard deviation (i.e., mean = 0, SD = 1).
}
