% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize.numeric}
\alias{standardize.numeric}
\title{Standardization}
\usage{
\method{standardize}{numeric}(x, robust = FALSE, ...)
}
\arguments{
\item{x}{Object.}

\item{robust}{If TRUE, the standardization will be based on \link{median} and \link{mad} instead of \link{mean} and \link{sd} (default).}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Standardize (scale and reduce, Z-score) X so that the values are expressed in terms of standard deviation (i.e., mean = 0, SD = 1).
}
