% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize}
\alias{standardize}
\title{Standardization}
\usage{
standardize(x, robust = FALSE, select = NULL, exclude = NULL, ...)
}
\arguments{
\item{x}{Object.}

\item{robust}{If TRUE, the standardization will be based on \link{median} and \link{mad} instead of \link{mean} and \link{sd} (default).}

\item{select}{For a data.frame, character or list of characters of column names to be
standardized.}

\item{exclude}{For a data.frame, character or list of characters of column names to be excluded
from standardization.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Standardize (scale and reduce, Z-score) X so that the values are expressed in terms of standard deviation (i.e., mean = 0, SD = 1).
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
