% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\name{table1}
\alias{table1}
\title{Create Table 1 (Sample Description)}
\usage{
table1(
  data,
  centrality = "mean",
  select = NULL,
  exclude = NULL,
  group_by = NULL,
  digits = 1,
  ...
)
}
\arguments{
\item{data}{A data frame for which descriptive statistics should be created.}

\item{centrality}{Character, indicates the statistics that should be calculated for numeric variables. May be \code{"mean"} (for mean and standard deviation) or \code{"median"} (for median and median absolute deviation) as summary.}

\item{select}{Character vector, with column names that should be included in the descriptive table.}

\item{exclude}{Character vector, with column names that should be excluded from the descriptive table.}

\item{group_by}{Character vector, indicating the column for possible grouping of the descriptive table.}

\item{digits}{Number of decimals.}
}
\description{
Create Table 1 (Sample Description).
}
\examples{
data(iris)
table1(iris[, 1:4])
table1(iris, select = c("Sepal.Length", "Petal.Length", "Species"))
table1(iris, group_by = "Species")
}
