% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_text.R
\name{text_fullstop}
\alias{text_fullstop}
\alias{text_concatenate}
\alias{text_remove}
\title{Convenient text formatting functionalities}
\usage{
text_fullstop(text)

text_concatenate(text, sep = ", ", last = " and ")

text_remove(text, pattern = "", ...)
}
\arguments{
\item{text}{A character string.}

\item{sep}{Separator.}

\item{last}{Last separator.}

\item{pattern}{Character strings.}

\item{...}{Other arguments to be passed to or from other functions.}
}
\value{
A character string.
}
\description{
Convenience functions to manipulate and format text.
}
\examples{
library(report)

# Add full stop if missing
text_fullstop(c("something", "something else."))

# Smart concatenation
text_concatenate(c("First", "Second", "Last"))

# Remove parts of string
text_remove(c("one!", "two", "three!"), "!")
}
