% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_initial.R
\name{text_initial}
\alias{text_initial}
\title{Initial model state textual reporting}
\usage{
text_initial(model, parameters, ...)
}
\arguments{
\item{model}{Object.}

\item{parameters}{Parameters table.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Convert initial state parameters table to text.
}
\seealso{
report
}
