% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{text_long}
\alias{text_long}
\alias{text_short}
\alias{table_long}
\alias{table_short}
\title{Access report components}
\usage{
text_long(r, ...)

text_short(r, ...)

table_long(r, ...)

table_short(r, ...)
}
\arguments{
\item{r}{Object of class \link{report}.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Access report components
}
