% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_model.R
\name{text_model}
\alias{text_model}
\title{Model description textual reporting}
\usage{
text_model(model, details = FALSE, ...)
}
\arguments{
\item{model}{Object.}

\item{details}{Add details.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Model textual description.
}
\seealso{
report
}
