% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_parameters.R
\name{text_parameters}
\alias{text_parameters}
\title{Parameters textual reporting}
\usage{
text_parameters(model, parameters, prefix = "  - ", ...)
}
\arguments{
\item{model}{Object.}

\item{parameters}{Parameters table.}

\item{prefix}{The bullet in front of each sentence.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Convert parameters table to text.
}
\seealso{
report
}
