% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{to_fulltext}
\alias{to_fulltext}
\title{Report printing}
\usage{
to_fulltext(x, full = TRUE, ...)
}
\arguments{
\item{x}{An object of class \link{report}}

\item{full}{Show the full report.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Report printing
}
