% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{to_table}
\alias{to_table}
\title{Report table display}
\usage{
to_table(x, full = FALSE, digits = NULL, ...)
}
\arguments{
\item{x}{An object of class \link{report}}

\item{full}{Show the full report.}

\item{digits}{Format the digits.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Report table display
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
