% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{to_text}
\alias{to_text}
\title{Report printing}
\usage{
to_text(x, full = FALSE, ...)
}
\arguments{
\item{x}{An object of class \link{report}}

\item{full}{Show the full report.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Report printing
}
