% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{to_values}
\alias{to_values}
\title{Report values}
\usage{
to_values(x, ...)
}
\arguments{
\item{x}{Object of class \link{report}.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Report values
}
