% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-reports.R
\name{view_report_ols}
\alias{view_report_ols}
\alias{view_report_blr}
\alias{view_report_rfm}
\alias{view_report_descriptr}
\title{Sample reports}
\usage{
view_report_ols()

view_report_blr()

view_report_rfm()

view_report_descriptr()
}
\description{
View sample reports.
}
\examples{
\dontrun{
view_report_ols()
view_report_blr()
view_report_rfm()
view_report_descriptr()
}

}
