% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{as.report}
\alias{as.report}
\alias{is.report}
\title{Create or test objects of class \link{report}.}
\usage{
as.report(x, ...)

is.report(x)
}
\arguments{
\item{x}{An arbitrary R object.}

\item{...}{Args to be saved as attributes.}
}
\value{
A report object or a \code{TRUE/FALSE} value.
}
\description{
Allows to create or test whether an object is of the \code{report} class.
}
