% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{as.report}
\alias{as.report}
\title{create objects of class \link{report}.}
\usage{
as.report(model)
}
\arguments{
\item{model}{An arbitrary R object.}
}
\description{
create objects of class \link{report}.
}
