% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{as.report}
\alias{as.report}
\alias{is.report}
\title{Create and test objects of class \link{report}.}
\usage{
as.report(model, ...)

is.report(model)
}
\arguments{
\item{model}{An arbitrary R object.}

\item{...}{Args to be saved as attributes.}
}
\description{
Create and test objects of class \link{report}.
}
