% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cite_easystats.R
\name{cite_easystats}
\alias{cite_easystats}
\title{Cite the easystats ecosystem}
\usage{
cite_easystats()
}
\value{
An object of class \code{cite_easystats} that can be printed, summarized (using \code{summary()}), or transformed into a table (using \code{as.data.frame()}).
}
\description{
A convenient function for those who wish to cite the easystats packages.
}
\examples{
cite_easystats()
summary(cite_easystats())
as.data.frame(cite_easystats())

}
