% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{.display.data.frame}
\alias{.display.data.frame}
\title{Dataframe Pretty Printing}
\usage{
\method{.display}{data.frame}(x, cat = TRUE, sep = " | ")
}
\description{
Dataframe Pretty Printing
}
\keyword{internal}
