% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_model.R
\name{format_model}
\alias{format_model}
\title{Model Name Formatting}
\usage{
format_model(model)
}
\arguments{
\item{model}{A statistical model.}
}
\description{
Model Name Formatting
}
\examples{
model <- lm(Sepal.Length ~ Species, data = iris)
format_model(model)

if (require("lme4")) {
model <- lme4::lmer(Sepal.Length ~ Sepal.Width + (1|Species), data = iris)
format_model(model)
}

}
