% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_text.R
\name{format_text}
\alias{format_text}
\title{Format text}
\usage{
format_text(x, sep = ", ", last = " and ", width = NULL)
}
\arguments{
\item{x}{Character vector.}

\item{sep}{Separator.}

\item{last}{Last separator.}

\item{width}{Positive integer giving the target column for wrapping lines in the output. Can be "auto".}
}
\description{
Format text
}
\examples{
format_text(c("A", "B", "C"))

x <- paste(rep("a very long string", 50), collapse = " ")
cat(format_text(x, width = 50))
}
