% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_text.R
\name{format_text_wrap}
\alias{format_text_wrap}
\title{Autowrap String}
\usage{
format_text_wrap(x, width = NULL)
}
\arguments{
\item{x}{Character vector.}

\item{width}{Positive integer giving the target column for wrapping lines in the output.}
}
\description{
Autowrap String
}
\examples{
x <- paste(rep("a very long string", 50), collapse = " ")
format_text_wrap(x, width = 50)
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
