% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_value.R
\name{format_value}
\alias{format_value}
\title{Value Formatting}
\usage{
format_value(x, digits = 2)
}
\arguments{
\item{x}{number.}

\item{digits}{number of significant digits.}
}
\description{
Value Formatting
}
\examples{
format_value(1.20)
format_value(1.2)
format_value(1.2012313)
format_value(0.0045)

}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
