% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_d.R
\name{interpret_d}
\alias{interpret_d}
\title{Standardized Difference (Cohen's d) Interpretation}
\usage{
interpret_d(d, rules = "cohen1988")
}
\arguments{
\item{d}{Value or vector of d values.}

\item{rules}{Can be "cohen1988" (default), "sawilowsky2009" or custom set of rules.}
}
\description{
Standardized Difference (Cohen's d) Interpretation
}
\examples{
interpret_d(.02)
interpret_d(c(.5, .02))
}
\references{
\itemize{
  \item Cohen, J. (1988). Statistical power analysis for the behavioural sciences.
  \item Sawilowsky, S. S. (2009). New effect size rules of thumb.
}
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
