% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_p.R
\name{interpret_p}
\alias{interpret_p}
\title{p-values Interpretation}
\usage{
interpret_p(p, rules = "default")
}
\arguments{
\item{p}{Value or vector of p-values.}

\item{rules}{Can be "default" or custom set of rules.}
}
\description{
p-values Interpretation
}
\examples{
interpret_p(.02)
interpret_p(c(.5, .02))
}
