% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.lm.R
\name{model_description.lm}
\alias{model_description.lm}
\title{Linear Models Description}
\usage{
\method{model_description}{lm}(model, effsize = "cohen1988", ...)
}
\arguments{
\item{model}{Object of class \link{lm}.}

\item{effsize}{Compute standardized parameters and interpret them using a set of rules. Can be "cohen1988" (default), "sawilowsky2009", NULL, or a custom set of \link{rules}.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Description for linear models.
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
