% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_table.R, R/model_table_regression.R
\name{model_table}
\alias{model_table}
\alias{model_table.default}
\alias{model_table.stanreg}
\title{Table Summary}
\usage{
model_table(model, ...)

\method{model_table}{default}(model, ...)

\method{model_table}{stanreg}(model, ci_method = "hdi", verbose = FALSE, ...)
}
\arguments{
\item{model}{A statistical model.}

\item{...}{Arguments passed to other methods, in particular \code{\link[parameters:model_parameters]{model_parameters()}}.}

\item{ci_method}{The type of index used for Credible Interval. Can be
\code{"HDI"} (default, see \code{\link[bayestestR:hdi]{hdi}}), \code{"ETI"}
(see \code{\link[bayestestR:eti]{eti}}) or \code{"SI"}
(see \code{\link[bayestestR:si]{si}}).}

\item{verbose}{Toggle warnings.}
}
\value{
A list with elements of class \code{report_table} (which are data
frame), containing a short and long version of the table output of the
model summary.
}
\description{
Create a reportable table (no pun intended) for your model.
}
\note{
Arguments in \code{...} are described in the \code{\link{report}} function help.
}
\examples{
model <- lm(Sepal.Length ~ Petal.Width, data = iris)
model_table(model)
summary(model_table(model))

model <- t.test(Sepal.Length ~ Species, data = iris[1:100, ])
model_table(model)
summary(model_table(model))
}
