% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.lm.R
\name{model_table}
\alias{model_table}
\title{TEST}
\usage{
model_table(model)
}
\arguments{
\item{model}{Object of class \link{lm}.}
}
\description{
TEST
}
\examples{
model <- lm(Sepal.Length ~ Petal.Length * Species, data = iris)
model_table(model)$table
}
