% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_model_methods.R
\name{model_table}
\alias{model_table}
\title{Model table}
\usage{
model_table(model, ...)
}
\arguments{
\item{model}{Statistical Model.}

\item{...}{Arguments passed to or from other methods (see \link{model_parameters} and \link{model_performance}).}
}
\description{
Model table. See the documentation for your object's class:
\itemize{
 \item{\link[=model_table.lm]{lm}}
 }
}
