% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_table.R, R/model_table_regression.R
\name{model_table}
\alias{model_table}
\alias{model_table.lm}
\title{Table Summary}
\usage{
model_table(model, ...)

\method{model_table}{lm}(model, ci = 0.95, bootstrap = FALSE, iterations = 500, ...)
}
\arguments{
\item{model}{A statistical model.}

\item{...}{Arguments passed to or from other methods. For instance, when \code{bootstrap = TRUE}, arguments like \code{ci_method} are passed down to \code{\link[bayestestR]{describe_posterior}}.}

\item{ci}{Confidence Interval (CI) level. Default to 0.95 (95\%).}

\item{bootstrap}{Should estimates be based on bootstrapped model? If \code{TRUE}, then arguments of \link[parameters:model_parameters.stanreg]{Bayesian regressions} apply (see also \code{\link[parameters:parameters_bootstrap]{parameters_bootstrap()}}).}

\item{iterations}{The number of bootstrap replicates. This only apply in the case of bootstrapped frequentist models.}
}
\description{
Create a reportable table (no pun intended) for your model.
}
\examples{
model <- lm(Sepal.Length ~ Petal.Width, data=iris)
model_table(model)
summary(model_table(model))

model <- t.test(Sepal.Length ~ Species, data=iris[1:100,])
model_table(model)
summary(model_table(model))

}
