% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_methods.R
\name{model_table}
\alias{model_table}
\title{Model table}
\usage{
model_table(model, ...)
}
\arguments{
\item{model}{Statistical Model.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Model table
}
