% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_table.R
\name{model_table}
\alias{model_table}
\title{Table Summary}
\usage{
model_table(model, ...)
}
\arguments{
\item{model}{A statistical model.}
}
\description{
Create a reportable table (no pun intended) for your model.
}
\examples{
model <- lm(Sepal.Length ~ Petal.Width, data=iris)
model_table(model)
summary(model_table(model))

model <- t.test(Sepal.Length ~ Species, data=iris[1:100,])
model_table(model)
summary(model_table(model))

}
