% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_text.R, R/model_text_regression.R
\name{model_text}
\alias{model_text}
\alias{model_text.default}
\title{Text summary}
\usage{
model_text(model, ...)

\method{model_text}{default}(model, ...)
}
\arguments{
\item{model}{A statistical model.}

\item{...}{Arguments passed to or from other methods. For instance, when \code{bootstrap = TRUE}, arguments like \code{ci_method} are passed down to \code{\link[bayestestR]{describe_posterior}}.}
}
\value{
A list with elements of class \code{report_text} (which are character
vectors), containing a short and long version of the textual output of the
model summary.
}
\description{
Create a reportable textual output for your model.
}
\note{
Arguments in \code{...} are described in the \code{\link{report}} function help.

Arguments in \code{...} are described in the \code{\link{report}} function help.
}
\examples{
model <- t.test(Sepal.Length ~ Species, data = iris[1:100, ])
model_text(model)
summary(model_text(model))
model <- lm(Sepal.Length ~ Petal.Length * Species, data = iris)
model_text(model)
summary(model_text(model))
}
