% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_text.R
\name{model_text}
\alias{model_text}
\title{Text summary}
\usage{
model_text(model, ...)
}
\arguments{
\item{model}{A statistical model.}
}
\description{
Create a reportable textual output for your model.
}
\examples{
model <- lm(Sepal.Length ~ Petal.Width, data=iris)
model_text(model)
summary(model_text(model))

model <- t.test(Sepal.Length ~ Species, data=iris[1:100,])
model_text(model)
summary(model_text(model))

}
