% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.lm.R
\name{model_values.lm}
\alias{model_values.lm}
\title{Linear Models Values}
\usage{
\method{model_values}{lm}(model, effsize = "cohen1988", ci = 0.95,
  performance_in_table = TRUE, ...)
}
\arguments{
\item{model}{Object of class \link{lm}.}

\item{effsize}{Compute standardized parameters and interpret them using a set of rules. Can be "cohen1988" (default), "sawilowsky2009", NULL, or a custom set of \link{rules}.}

\item{ci}{Confidence Interval (CI) level. Default to 0.95 (95\%).}

\item{performance_in_table}{Add performance metrics on table.}

\item{...}{Arguments passed to or from other methods (see \link{model_parameters} and \link{model_performance}).}
}
\description{
Extract all values of linear models.
}
