% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_methods.R
\name{model_values}
\alias{model_values}
\title{Model Values}
\usage{
model_values(model, ...)
}
\arguments{
\item{model}{Statistical Model.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Model Values
}
