% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{print.report}
\alias{print.report}
\title{Report printing}
\usage{
\method{print}{report}(x, full = FALSE, ...)
}
\arguments{
\item{x}{Object of class \link{report}}

\item{full}{Show the full report.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Report printing
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
