% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{correlation}
\alias{estimate_means}
\alias{estimate_contrasts}
\alias{estimate_slopes}
\alias{estimate_smooth}
\alias{data_grid}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{correlation}{\code{\link[correlation]{correlation}}, \code{\link[correlation]{correlation}}}

  \item{estimate}{\code{\link[estimate]{estimate_means}}, \code{\link[estimate]{estimate_contrasts}}, \code{\link[estimate]{estimate_slopes}}, \code{\link[estimate]{estimate_smooth}}, \code{\link[estimate]{data_grid}}}
}}

