% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{data_rename}
\alias{data_addprefix}
\alias{data_addsuffix}
\alias{data_findcols}
\alias{data_remove}
\alias{data_reorder}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{datawizard}{\code{\link[datawizard:data_rename]{data_addprefix}}, \code{\link[datawizard:data_rename]{data_addsuffix}}, \code{\link[datawizard:data_rename]{data_findcols}}, \code{\link[datawizard:data_rename]{data_remove}}, \code{\link[datawizard]{data_rename}}, \code{\link[datawizard:data_rename]{data_reorder}}}
}}

