% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_table.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{display}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{insight}{\code{\link[insight]{display}}}
}}

