% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{model_parameters}
\alias{model_performance}
\alias{model_info}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{insight}{\code{\link[insight]{model_info}}}

  \item{parameters}{\code{\link[parameters]{model_parameters}}}

  \item{performance}{\code{\link[performance]{model_performance}}}
}}

