% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rename_if_possible.R
\name{rename_if_possible}
\alias{rename_if_possible}
\title{Rename a Variable}
\usage{
rename_if_possible(x, pattern, replacement)
}
\arguments{
\item{x}{Dataframe.}

\item{pattern}{Variable to replace.}

\item{replacement}{New variable name.}
}
\description{
Rename a variable in a dataframe if applicable.
}
\examples{
rename_if_possible(iris, "Sepal.Length", "length")
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
