% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dataframe.R
\name{rename_if_possible}
\alias{rename_if_possible}
\alias{remove_if_possible}
\alias{reorder_if_possible}
\title{Safe Dataframe Modification}
\usage{
rename_if_possible(x, pattern, replacement)

remove_if_possible(x, pattern)

reorder_if_possible(x, cols)
}
\arguments{
\item{x}{Dataframe.}

\item{pattern}{Variable to replace.}

\item{replacement}{New variable name.}

\item{cols}{Order of columns.}
}
\description{
Rename/remove a variable in a dataframe if exists.
}
\examples{
rename_if_possible(iris, "Sepal.Length", "length")
rename_if_possible(iris, "FakeCol", "length")
remove_if_possible(iris, "Sepal.Length")
reorder_if_possible(iris, c("Species", "Sepal.Length"))
}
