% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rename_if_possible.R
\name{rename_if_possible}
\alias{rename_if_possible}
\alias{remove_if_possible}
\title{Safe Dataframe Modification}
\usage{
rename_if_possible(x, pattern, replacement)

remove_if_possible(x, pattern)
}
\arguments{
\item{x}{Dataframe.}

\item{pattern}{Variable to replace.}

\item{replacement}{New variable name.}
}
\description{
Rename/remove a variable in a dataframe if exists.
}
\examples{
rename_if_possible(iris, "Sepal.Length", "length")
rename_if_possible(iris, "FakeCol", "length")
remove_if_possible(iris, "Sepal.Length")
}
