% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-package.R
\docType{package}
\name{report-package}
\alias{report}
\alias{report-package}
\title{report: Automated Results Reporting as a Practical Tool to Improve
Reproducibility and Methodological Best Practices Adoption}
\description{
\strong{report}’s primary goal is to bridge the gap between R’s output and
the formatted results contained in your manuscript. It automatically produces
reports of models and data frames according to \strong{best practices} guidelines
(e.g., \href{https://apastyle.apa.org/}{APA}’s style), ensuring \strong{standardization}
and \strong{quality} in results reporting.
}
\details{
\code{report-package}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://easystats.github.io/report/}
  \item Report bugs at \url{https://github.com/easystats/report/issues}
}

}
\author{
\strong{Maintainer}: Dominique Makowski \email{dom.makowski@gmail.com} (\href{https://orcid.org/0000-0001-5375-9967}{ORCID}) (@Dom_Makowski)

Authors:
\itemize{
  \item Daniel Lüdecke \email{d.luedecke@uke.de} (\href{https://orcid.org/0000-0002-8895-3206}{ORCID}) (@strengejacke)
  \item Mattan S. Ben-Shachar \email{matanshm@post.bgu.ac.il} (\href{https://orcid.org/0000-0002-4287-4801}{ORCID}) (@mattansb)
  \item Indrajeet Patil \email{patilindrajeet.science@gmail.com} (\href{https://orcid.org/0000-0003-1995-6531}{ORCID}) (@patilindrajeets)
  \item Brenton M. Wiernik \email{brenton@wiernik.org} (\href{https://orcid.org/0000-0001-9560-6336}{ORCID}) (@bmwiernik)
  \item Rémi Thériault \email{remi.theriault@mail.mcgill.ca} (\href{https://orcid.org/0000-0003-4315-6788}{ORCID}) (@rempsyc)
}

Other contributors:
\itemize{
  \item Rudolf Siegel \email{mutlusun@users.noreply.github.com} (\href{https://orcid.org/0000-0002-6021-804X}{ORCID}) [contributor]
  \item Rémi Thériault \email{remi.theriault@mail.mcgill.ca} (\href{https://orcid.org/0000-0003-4315-6788}{ORCID}) (@rempsyc) [contributor]
}

}
\keyword{internal}
