% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.aov.R
\name{report.aov}
\alias{report.aov}
\title{ANOVAs Report}
\usage{
\method{report}{aov}(model, omega_squared = "partial",
  effsize = "field2013", ...)
}
\arguments{
\item{model}{Object of class \code{aov}, \code{anova} or \code{aovlist}.}

\item{omega_squared}{Compute omega squared as indices of effect size. Can be \code{NULL}, "partial" (default) or "raw" for non-partial indices.}

\item{effsize}{Effect size interpretation set of rules.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Create a report of an ANOVA.
}
\examples{
data <- iris
data$Cat1 <- rep(c("X", "X", "Y"), length.out = nrow(data))
data$Cat2 <- rep(c("A", "B"), length.out = nrow(data))
model <- aov(Sepal.Length ~ Species * Cat1 * Cat2, data=data)
report(model)


\dontrun{
report(circus::aov_1)
report(circus::anova_1)
report(circus::aovlist_1)
}
}
\seealso{
report
}
