% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.aov.R
\name{report.aov}
\alias{report.aov}
\title{ANOVAs Report}
\usage{
\method{report}{aov}(x, ...)
}
\arguments{
\item{x}{Object of class \code{aov}, \code{anova} or \code{aovlist}.}

\item{...}{Arguments passed to or from other methods.}
}
\value{
A list-object of class \code{report}, which contains further list-objects
with a short and long description of the model summary, as well as a short
and long table of parameters and fit indices.
}
\description{
Create a report of an ANOVA.
}
\examples{
data <- iris
data$Cat1 <- rep(c("A", "B"), length.out = nrow(data))

model <- aov(Sepal.Length ~ Species * Cat1, data = data)
r <- report(model)
r
summary(r)
as.data.frame(r)
summary(as.data.frame(r))
}
\seealso{
Aspects of reports:
\itemize{
  \item \code{\link{report_table}}
  \item \code{\link{report_parameters}}
  \item \code{\link{report_statistics}}
  \item \code{\link{report_effectsize}}
  \item \code{\link{report_model}}
  \item \code{\link{report_performance}}
  \item \code{\link{report_info}}
  \item \code{\link{report_text}}
}
Methods:
\itemize{
  \item \code{\link{as.report}}
}
Specific reports:
\itemize{
  \item \code{\link{report_system}}
  \item \code{\link{report_packages}}
  \item \code{\link{report_participants}}
  \item \code{\link{report_sample}}
}
}
