% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.aov.R
\name{report.aov}
\alias{report.aov}
\title{ANOVAs Report}
\usage{
\method{report}{aov}(model, effsize = "field2013",
  omega_squared = NULL, eta_squared = NULL, epsilon_squared = NULL,
  ...)
}
\arguments{
\item{model}{Object of class \code{aov}, \code{anova} or \code{aovlist}.}

\item{effsize}{Effect size interpretation set of rules.}

\item{omega_squared}{Compute omega squared as index of effect size. Can be "partial" (adjusted for effect size) or "raw".}

\item{eta_squared}{Compute eta squared as index of effect size. Can be "partial" (adjusted for effect size) or "raw".}

\item{epsilon_squared}{Compute epsilon squared as index of effect size.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Create a report of an ANOVA.
}
\examples{
data <- iris
data$Cat1 <- rep(c("X", "X", "Y"), length.out = nrow(data))
data$Cat2 <- rep(c("A", "B"), length.out = nrow(data))
model <- aov(Sepal.Length ~ Species * Cat1 * Cat2, data = data)
report(model, omega_squared = "partial")
}
\seealso{
report
}
