% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.bayesfactor_models.R
\name{report.bayesfactor_models}
\alias{report.bayesfactor_models}
\alias{report.bayesfactor_inclusion}
\title{Report Models' Bayes Factor}
\usage{
\method{report}{bayesfactor_models}(x, interpretation = "jeffreys1961", ...)

\method{report}{bayesfactor_inclusion}(x, interpretation = "jeffreys1961", ...)
}
\arguments{
\item{x}{Object of class \code{bayesfactor_inclusion}.}

\item{interpretation}{Effect size interpretation set of rules (see \link[effectsize]{interpret_bf}).}

\item{...}{Arguments passed to or from other methods.}
}
\value{
A list-object of class \code{report}, which contains further list-objects
with a short and long description of the model summary, as well as a short
and long table of parameters and fit indices.
}
\description{
Create a report of Bayes factors for model comparison.
}
\examples{
library(report)

mo0 <- lm(Sepal.Length ~ 1, data = iris)
mo1 <- lm(Sepal.Length ~ Species, data = iris)
mo2 <- lm(Sepal.Length ~ Species + Petal.Length, data = iris)
mo3 <- lm(Sepal.Length ~ Species * Petal.Length, data = iris)

if (require("bayestestR")) {
  # Bayes factor - models
  BFmodels <- bayesfactor_models(mo1, mo2, mo3, denominator = mo0)

  r <- report(BFmodels)
  r
  as.data.frame(r)

  # Bayes factor - inclusion
  inc_bf <- bayesfactor_inclusion(BFmodels, prior_odds = c(1, 2, 3), match_models = TRUE)

  r <- report(inc_bf)
  r
  as.data.frame(r)
}
}
\seealso{
Aspects of reports:
\itemize{
  \item \code{\link{report_table}}
  \item \code{\link{report_parameters}}
  \item \code{\link{report_statistics}}
  \item \code{\link{report_effectsize}}
  \item \code{\link{report_model}}
  \item \code{\link{report_performance}}
  \item \code{\link{report_info}}
  \item \code{\link{report_text}}
}
Methods:
\itemize{
  \item \code{\link{as.report}}
}
Specific reports:
\itemize{
  \item \code{\link{report_system}}
  \item \code{\link{report_packages}}
  \item \code{\link{report_participants}}
  \item \code{\link{report_sample}}
}
}
