% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.BFBayesFactor.R
\name{report.BFBayesFactor}
\alias{report.BFBayesFactor}
\alias{report_statistics.BFBayesFactor}
\title{Reporting \code{BFBayesFactor} objects from the \code{BayesFactor} package}
\usage{
\method{report}{BFBayesFactor}(x, h0 = "H0", h1 = "H1", ...)

\method{report_statistics}{BFBayesFactor}(x, table = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{BFBayesFactor}.}

\item{h0, h1}{Names of the null and alternative hypotheses.}
}
\description{
Interpretation of the Bayes factor output from the \code{BayesFactor} package.
}
\examples{
\dontshow{if (requireNamespace("BayesFactor", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\donttest{
library(BayesFactor)

rez <- BayesFactor::ttestBF(iris$Sepal.Width, iris$Sepal.Length)
report_statistics(rez)
report(rez, h0="the null hypothesis", h1="the alternative")

rez <- BayesFactor::correlationBF(iris$Sepal.Width, iris$Sepal.Length)
report(rez)
}
\dontshow{\}) # examplesIf}
}
