% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.character.R
\name{report.character}
\alias{report.character}
\title{Character Vector Report}
\usage{
\method{report}{character}(x, n_characters = 3,
  levels_percentage = TRUE, missing_percentage = FALSE, ...)
}
\arguments{
\item{x}{a character vector.}

\item{n_characters}{number of different character entries to show. Can be "all".}

\item{levels_percentage}{show characters entries by percentage (default) or number.}

\item{missing_percentage}{show missings by number (default) or percentage}

\item{...}{arguments passed to or from other methods.}
}
\description{
Create a report of a character vector.
}
\examples{
x <- c("A", "B", "C", "A", "B", "B", "D", "E", "B", "D", "A")
report(x)
report(x, n_characters = 2, levels_percentage = FALSE, missing_percentage = TRUE)
to_fulltext(report(x))
to_table(report(x, n_characters = "all"))
to_fulltable(report(x))
}
\seealso{
report
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
