% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.correlation.R
\name{report.correlation}
\alias{report.correlation}
\title{Correlation Report}
\usage{
\method{report}{correlation}(model, effsize = "cohen1988",
  stars = TRUE, reorder = TRUE, ...)
}
\arguments{
\item{model}{Object of class correlation}

\item{effsize}{Effect size interpretation set of rules.}

\item{stars}{Add significance stars in table. For frequentist correlations: \*p < .05, \*\*p < .01, \*\*\*p < .001. For Bayesian correlations: \*\*\*BF > 30, \*\*BF > 10, \*BF > 3.}

\item{reorder}{Should the tables be reordered based on correlation pattern (currently only works with square matrices)?}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Create a report of a correlation object.
}
\examples{
model <- correlation(iris)
report(model)
}
\seealso{
report
}
