% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.data.frame.R, R/report.numeric.R
\name{report.data.frame}
\alias{report.data.frame}
\alias{report.numeric}
\title{Dataframe Report}
\usage{
\method{report}{data.frame}(
  model,
  median = FALSE,
  centrality = TRUE,
  dispersion = TRUE,
  range = TRUE,
  distribution = FALSE,
  levels_percentage = FALSE,
  n_entries = 3,
  missing_percentage = FALSE,
  ...
)

\method{report}{numeric}(
  model,
  median = FALSE,
  centrality = TRUE,
  dispersion = TRUE,
  range = TRUE,
  distribution = FALSE,
  missing_percentage = FALSE,
  ...
)
}
\arguments{
\item{model}{A data.frame or a vector.}

\item{median}{Show \link{mean} and \link{sd} (default) or \link{median} and \link{mad}.}

\item{centrality}{Show index of centrality (\link{mean} or \link{median}).}

\item{dispersion}{Show index of dispersion (\link{sd} or \link{mad}).}

\item{range}{Show range.}

\item{distribution}{Returns Kurtosis and Skewness in table.}

\item{levels_percentage}{Show characters entries and factor levels by number (default) or percentage.}

\item{n_entries}{Number of different character entries to show. Can be "all".}

\item{missing_percentage}{Show missings by number (default) or percentage.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Create a report of a dataframe.
}
\examples{
library(report)

report(iris)
r <- report(iris, median = TRUE, dispersion = FALSE, distribution = TRUE, missing_percentage = TRUE)

x <- rnorm(1000)
report(x)
report(x, median = TRUE, missing_percentage = TRUE, distribution = TRUE)

}
\seealso{
report

report
}
