% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.character.R, R/report.data.frame.R,
%   R/report.numeric.R
\name{report.character}
\alias{report.character}
\alias{report.data.frame}
\alias{report.numeric}
\title{Data frame Report}
\usage{
\method{report}{character}(
  x,
  n_entries = 3,
  levels_percentage = FALSE,
  missing_percentage = FALSE,
  ...
)

\method{report}{data.frame}(
  x,
  centrality = "mean",
  dispersion = TRUE,
  range = TRUE,
  distribution = FALSE,
  levels_percentage = FALSE,
  n_entries = 3,
  missing_percentage = FALSE,
  ...
)

\method{report}{numeric}(
  x,
  centrality = "mean",
  dispersion = TRUE,
  range = TRUE,
  distribution = FALSE,
  missing_percentage = FALSE,
  digits = 2,
  ...
)
}
\arguments{
\item{x}{The R object that you want to report (see list of of supported objects above).}

\item{n_entries}{Number of different character entries to show. Can be "all".}

\item{levels_percentage}{Show characters entries and factor levels by number (default) or percentage.}

\item{missing_percentage}{Show missing by number (default) or percentage.}

\item{...}{Arguments passed to or from other methods.}

\item{centrality}{Character vector, indicating the index of centrality (either \code{"mean"} or \code{"median"}).}

\item{dispersion}{Show index of dispersion (\link{sd} if \code{centrality = "mean"}, or \link{mad} if \code{centrality = "median"}).}

\item{range}{Show range.}

\item{distribution}{Returns kurtosis and skewness in table.}

\item{digits}{Number of significant digits.}
}
\description{
Create a report of a data frame.
}
\examples{
library(report)

r <- report(iris,
  centrality = "median", dispersion = FALSE,
  distribution = TRUE, missing_percentage = TRUE
)
r
summary(r)
as.data.frame(r)
summary(as.data.frame(r))

if (require("dplyr")) {
  r <- iris \%>\%
    dplyr::group_by(Species) \%>\%
    report()
  r
  summary(r)
  as.data.frame(r)
  summary(as.data.frame(r))
}
}
