% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.data.frame.R
\name{report.data.frame}
\alias{report.data.frame}
\title{Dataframe Report}
\usage{
\method{report}{data.frame}(x, median = FALSE, dispersion = TRUE,
  range = TRUE, levels_percentage = FALSE, n_characters = 3,
  missing_percentage = FALSE, ...)
}
\arguments{
\item{x}{Numeric vector.}

\item{median}{Show \link{mean} and \link{sd} (default) or \link{median} and \link{mad}.}

\item{dispersion}{Show dispersion (\link{sd} or \link{mad}).}

\item{range}{Show range.}

\item{levels_percentage}{Show characters entries and factor levels by number (default) or percentage.}

\item{n_characters}{Number of different character entries to show. Can be "all".}

\item{missing_percentage}{Show missings by number (default) or percentage.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Create a report of a dataframe.
}
\examples{
x <- iris
report(x)
report(x, median = TRUE, dispersion = TRUE, range = TRUE, missing_percentage = TRUE)
to_fulltext(report(x))
to_table(report(x))
to_fulltable(report(x))
}
\seealso{
report
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
