% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.data.frame.R
\name{report.data.frame}
\alias{report.data.frame}
\alias{report.character}
\alias{report.factor}
\alias{report.numeric}
\title{Dataframe Report}
\usage{
\method{report}{data.frame}(model, median = FALSE, centrality = TRUE,
  dispersion = TRUE, range = TRUE, distribution = FALSE,
  levels_percentage = FALSE, n_characters = 3,
  missing_percentage = FALSE, ...)

\method{report}{character}(model, n_characters = 3,
  levels_percentage = FALSE, missing_percentage = FALSE, ...)

\method{report}{factor}(model, levels_percentage = FALSE, ...)

\method{report}{numeric}(model, median = FALSE, centrality = TRUE,
  dispersion = TRUE, range = TRUE, distribution = FALSE,
  missing_percentage = FALSE, ...)
}
\arguments{
\item{model}{A data.frame or a vector.}

\item{median}{Show \link{mean} and \link{sd} (default) or \link{median} and \link{mad}.}

\item{centrality}{Show index of centrality (\link{mean} or \link{median}).}

\item{dispersion}{Show index of dispersion (\link{sd} or \link{mad}).}

\item{range}{Show range.}

\item{distribution}{Returns Kurtosis and Skewness in table.}

\item{levels_percentage}{Show characters entries and factor levels by number (default) or percentage.}

\item{n_characters}{Number of different character entries to show. Can be "all".}

\item{missing_percentage}{Show missings by number (default) or percentage.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Create a report of a dataframe.
}
\examples{
library(report)

report(iris)
r <- report(iris, median = TRUE, dispersion = FALSE, distribution = TRUE, missing_percentage = TRUE)
to_fulltext(r)
to_table(r)
to_fulltable(r)
x <- c("A", "B", "C", "A", "B", "B", "D", "E", "B", "D", "A")
report(x)
report(x, n_characters = 2, levels_percentage = TRUE, missing_percentage = TRUE)
to_fulltext(report(x))
to_table(report(x, n_characters = "all"))
to_fulltable(report(x))
x <- factor(rep(c("A", "B", "C"), 10))
report(x)
report(x, levels_percentage = TRUE, missing_percentage = TRUE)
to_fulltext(report(x))
to_table(report(x))
to_fulltable(report(x))
\dontrun{
x <- rnorm(1000)
report(x)
report(x, median = TRUE, missing_percentage = TRUE, distribution = TRUE)
to_fulltext(report(x))
to_table(report(x))
to_fulltable(report(x))
}
}
\seealso{
report

report

report

report
}
