% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.estimate.R
\name{report.estimate_contrasts}
\alias{report.estimate_contrasts}
\alias{report.estimate_means}
\alias{report.estimate_slopes}
\alias{report.estimate_smooth}
\title{Estimated Means, Contrasts, Slopes and Smooth Report}
\usage{
\method{report}{estimate_contrasts}(model, effsize = "funder2019", ...)

\method{report}{estimate_means}(model, effsize = "funder2019", ...)

\method{report}{estimate_slopes}(model, effsize = "funder2019", ...)

\method{report}{estimate_smooth}(model, ...)
}
\arguments{
\item{model}{Object of class \code{estimate}.}

\item{effsize}{\href{https://easystats.github.io/report/articles/interpret_metrics.html}{Interpret the standardized parameters} using a set of rules. Default corresponds to "funder2019" for linear models and "chen2010" for logistic models.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Create a report of an object from the \code{estimate} package.
}
\examples{
library(estimate)
library(report)
library(lme4)

data <- iris
data$Group <- ifelse(data$Sepal.Width > 3, "A","B")

model <- lm(Petal.Width ~ Species * Group, data = data)
report(estimate_contrasts(model))
\dontrun{
library(rstanarm)

model <- stan_glm(Sepal.Width ~ Species * Petal.Width, data = iris)
report(estimate_contrasts(model))
report(estimate_contrasts(model, fixed = "Petal.Width"))
report(estimate_contrasts(model, modulate = "Petal.Width", length = 2))
}

\dontrun{
library(estimate)
library(rstanarm)

model <- stan_glm(Sepal.Width ~ Species * Petal.Width, data = iris)
report(estimate_means(model))
report(estimate_means(model, modulate = "Petal.Width"))
}
\dontrun{
library(rstanarm)
model <- stan_glm(Sepal.Width ~ Species * Petal.Width, data = iris)
report(estimate_slopes(model))
}
\dontrun{
library(rstanarm)
model <- stan_glm(Sepal.Width ~ poly(Petal.Length, 2), data = iris)
report(estimate_smooth(model))
}
}
\seealso{
report
}
