% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.factor.R
\name{report.factor}
\alias{report.factor}
\title{Factor (Categorical Vector) Report}
\usage{
\method{report}{factor}(model, levels_percentage = FALSE,
  missing_percentage = FALSE, ...)
}
\arguments{
\item{model}{Categorical vector.}

\item{levels_percentage}{Show factor levels by number (default) or percentage.}

\item{missing_percentage}{Show missings by number (default) or percentage.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Create a report of a categorical vector.
}
\examples{
x <- factor(rep(c("A", "B", "C"), 10))
report(x)
report(x, levels_percentage = TRUE, missing_percentage = TRUE)
to_fulltext(report(x))
to_table(report(x))
to_fulltable(report(x))
}
\seealso{
report
}
