% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.htest.R
\name{report.htest}
\alias{report.htest}
\title{Report of h-tests (Correlation, t-test...)}
\usage{
\method{report}{htest}(x, ...)
}
\arguments{
\item{x}{Object of class htest.}

\item{...}{Arguments passed to or from other methods.}
}
\value{
A list-object of class \code{report}, which contains further list-objects
with a short and long description of the model summary, as well as a short
and long table of parameters and fit indices.
}
\description{
Create a report of an h-test object (\code{t.test()}, \code{cor.test()}).
}
\examples{
report(cor.test(iris$Sepal.Width, iris$Sepal.Length, method = "spearman"))
report(cor.test(iris$Sepal.Width, iris$Sepal.Length, method = "pearson"))
report(t.test(iris$Sepal.Width, iris$Sepal.Length))
report(t.test(iris$Sepal.Width, iris$Sepal.Length, var.equal = TRUE))
report(t.test(mtcars$mpg ~ mtcars$vs))
report(t.test(iris$Sepal.Width, mu = 1))
}
\seealso{
Aspects of reports:
\itemize{
  \item \code{\link{report_table}}
  \item \code{\link{report_parameters}}
  \item \code{\link{report_statistics}}
  \item \code{\link{report_effectsize}}
  \item \code{\link{report_model}}
  \item \code{\link{report_text}}
}
Methods:
\itemize{
  \item \code{\link{as.report}}
}
Specific reports:
\itemize{
  \item \code{\link{report_system}}
  \item \code{\link{report_packages}}
  \item \code{\link{report_participants}}
  \item \code{\link{report_sample}}
}
}
