% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_text.htest.R, R/report.htest.R
\name{model_text.htest}
\alias{model_text.htest}
\alias{report.htest}
\title{h-test (Correlation, t-test...) Report}
\usage{
\method{model_text}{htest}(model, interpretation = "funder2019", ...)

\method{report}{htest}(model, interpretation = "funder2019", ...)
}
\arguments{
\item{model}{Object of class htest.}

\item{interpretation}{Effect size interpretation set of rules (see \link[effectsize]{interpret_d} and \link[effectsize]{interpret_r}).}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Create a report of an h-test object.
}
\examples{
report(cor.test(iris$Sepal.Width, iris$Sepal.Length, method = "spearman"))
report(cor.test(iris$Sepal.Width, iris$Sepal.Length, method = "pearson"))
report(t.test(iris$Sepal.Width, iris$Sepal.Length))
report(t.test(iris$Sepal.Width, iris$Sepal.Length, var.equal = TRUE))
report(t.test(mtcars$mpg ~ mtcars$vs))
report(t.test(iris$Sepal.Width, mu = 1))
}
\seealso{
report

report
}
