% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.htest.R
\name{report.htest}
\alias{report.htest}
\alias{report_effectsize.htest}
\alias{report_table.htest}
\alias{report_statistics.htest}
\alias{report_parameters.htest}
\alias{report_model.htest}
\alias{report_info.htest}
\alias{report_text.htest}
\title{Reports of h-tests (Correlation, t-test...)}
\usage{
\method{report}{htest}(x, ...)

\method{report_effectsize}{htest}(x, ...)

\method{report_table}{htest}(x, ...)

\method{report_statistics}{htest}(x, table = NULL, ...)

\method{report_parameters}{htest}(x, table = NULL, ...)

\method{report_model}{htest}(x, table = NULL, ...)

\method{report_info}{htest}(x, effectsize = NULL, ...)

\method{report_text}{htest}(x, table = NULL, ...)
}
\arguments{
\item{x}{Object of class htest.}

\item{...}{Arguments passed to or from other methods.}

\item{table}{Provide the output of  \code{report_table()} to avoid its re-computation.}

\item{effectsize}{Provide the output of \code{report_effectsize()} to avoid its re-computation.}
}
\value{
A list-object of class \code{report}, which contains further list-objects
with a short and long description of the model summary, as well as a short
and long table of parameters and fit indices.
}
\description{
Create a report of an h-test object (\code{t.test()}, \code{cor.test()}).
}
\examples{
report(cor.test(iris$Sepal.Width, iris$Sepal.Length, method = "spearman"))
report(cor.test(iris$Sepal.Width, iris$Sepal.Length, method = "pearson"))
report(t.test(iris$Sepal.Width, iris$Sepal.Length))
report(t.test(iris$Sepal.Width, iris$Sepal.Length, var.equal = TRUE))
report(t.test(mtcars$mpg ~ mtcars$vs))
report(t.test(mtcars$mpg, mtcars$vs, paired=TRUE))
report(t.test(iris$Sepal.Width, mu = 1))
}
\seealso{
Specific components of reports (especially for stats models):
\itemize{
  \item \code{\link{report_table}}
  \item \code{\link{report_parameters}}
  \item \code{\link{report_statistics}}
  \item \code{\link{report_effectsize}}
  \item \code{\link{report_model}}
  \item \code{\link{report_random}}
  \item \code{\link{report_performance}}
  \item \code{\link{report_info}}
  \item \code{\link{report_text}}
}
Other types of reports:
\itemize{
  \item \code{\link{report_system}}
  \item \code{\link{report_packages}}
  \item \code{\link{report_participants}}
  \item \code{\link{report_sample}}
}
Methods:
\itemize{
  \item \code{\link{as.report}}
}
Template file for supporting new models:
\itemize{
  \item \code{\link{report.default}}
}
}
