% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.lavaan.R
\name{report.lavaan}
\alias{report.lavaan}
\title{SEM / CFA lavaan Models Report}
\usage{
\method{report}{lavaan}(model, effsize = "funder2019", ci = 0.95,
  standardize = TRUE, performance_metrics = "all", ...)
}
\arguments{
\item{model}{Model object.}

\item{effsize}{\href{https://easystats.github.io/report/articles/interpret_metrics.html}{Interpret the standardized parameters} using a set of rules. Default corresponds to "funder2019" for linear models and "chen2010" for logistic models.}

\item{ci}{Confidence Interval (CI) level. Default to 0.95 (95\%).}

\item{standardize}{Add standardized parameters. Can be \code{FALSE} or a character indicating the standardization method (see \code{\link[parameters:parameters_standardize]{parameters_standardize()}}), such as \code{"refit"}, \code{"2sd"}, \code{"smart"} or \code{"classic"}. The two formers are based on model refitting using a standardized version of data. It is the most accurate, altough computationnaly heavy (as it must re-fit a second model). The "smart" and "classic" are post-hoc methods, fast, but inaccurate (especially if the model includes interactions).}

\item{performance_metrics}{See \code{\link[performance:model_performance.lm]{model_performance}}.}

\item{...}{Arguments passed to or from other methods (e.g., to \code{\link[parameters:standardize.lm]{standardize()}}).}
}
\description{
Create a report of a lavaan model.
}
\examples{
library(report)
library(lavaan)

structure <- " visual  =~ x1 + x2 + x3
               textual =~ x4 + x5 + x6
               speed   =~ x7 + x8 + x9 "

model <- lavaan::cfa(structure, data = HolzingerSwineford1939)
report(model)
}
