% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.glm.R
\name{report.lm}
\alias{report.lm}
\title{(General) Linear Models Report}
\usage{
\method{report}{lm}(model, effsize = "default", ci = 0.95,
  standardize = "refit", standardize_robust = FALSE,
  bootstrap = FALSE, iterations = 500, performance_metrics = "all",
  ...)
}
\arguments{
\item{effsize}{\href{https://easystats.github.io/report/articles/interpret_metrics.html}{Interpret the standardized parameters} using a set of rules. Default corresponds to "funder2019" for linear models and "chen2010" for logistic models.}

\item{performance_metrics}{See \code{\link[performance:model_performance.lm]{model_performance}}.}
}
\description{
Create a report of a (general) linear model.
}
\examples{
library(report)

model <- lm(Sepal.Length ~ Petal.Length * Species, data = iris)
r <- report(model)
to_text(r)
to_fulltext(r)
to_table(r)
to_fulltable(r)


model <- glm(vs ~ disp, data = mtcars, family = "binomial")
r <- report(model)
to_text(r)
to_fulltext(r)
to_table(r)
to_fulltable(r)
}
