% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.lm.R
\name{report.lm}
\alias{report.lm}
\title{Linear Models Report}
\usage{
\method{report}{lm}(x, CI = 95, effsize = "cohen1988", data = NULL,
  ...)
}
\arguments{
\item{x}{Object of class \link{lm}.}

\item{CI}{Confidence Interval (CI) level. Default to 95\%.}

\item{effsize}{Compute standardized parameters and interpret them using a set of rules. Can be "cohen1988" (default), "sawilowsky2009", NULL, or a custom set of \link{rules}.}

\item{data}{Dataframe used to fit the model. Can be useful to compute standardized parameters.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Create a report of a linear model.
}
\examples{
model <- lm(Sepal.Length ~ Petal.Length * Species, data = iris)
r <- report(model)
to_text(r)
to_fulltext(r)
to_table(r)
to_fulltable(r)
}
