% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.lm.R
\name{report.lm}
\alias{report.lm}
\title{(General) Linear Models Report}
\usage{
\method{report}{lm}(x, include_effectsize = TRUE, effectsize_method = "refit", ...)
}
\arguments{
\item{x}{Object of class \code{lm} or \code{glm}.}

\item{include_effectsize}{Include effect size information.}

\item{effectsize_method}{See documentation for \code{\link[effectsize:effectsize]{effectsize::effectsize()}}.}

\item{...}{Arguments passed to or from other methods.}
}
\value{
A list-object of class \code{report}, which contains further list-objects
with a short and long description of the model summary, as well as a short
and long table of parameters and fit indices.
}
\description{
Create a report of a (general) linear model (i.e., a regression fitted using \code{lm()} or \code{glm()}.
}
\examples{
library(report)

# Linear models
model <- lm(Sepal.Length ~ Petal.Length * Species, data = iris)
r <- report(model)
r
summary(r)
as.data.frame(r)
summary(as.data.frame(r))

# Logistic models
model <- glm(vs ~ disp, data = mtcars, family = "binomial")
r <- report(model)
r
summary(r)
as.data.frame(r)
summary(as.data.frame(r))

# Mixed models
if(require("lme4")){
  model <- lme4::lmer(Sepal.Length ~ Petal.Length + (1 | Species), data = iris)
  r <- report(model)
  r
  summary(r)
  as.data.frame(r)
  summary(as.data.frame(r))
}
}
\seealso{
Aspects of reports (especially for stats models):
\itemize{
  \item \code{\link{report_table}}
  \item \code{\link{report_parameters}}
  \item \code{\link{report_statistics}}
  \item \code{\link{report_effectsize}}
  \item \code{\link{report_model}}
  \item \code{\link{report_random}}
  \item \code{\link{report_performance}}
  \item \code{\link{report_info}}
  \item \code{\link{report_text}}
}
Specific reports:
\itemize{
  \item \code{\link{report_system}}
  \item \code{\link{report_packages}}
  \item \code{\link{report_participants}}
  \item \code{\link{report_sample}}
}
Methods:
\itemize{
  \item \code{\link{as.report}}
}
}
