% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.lm.R, R/report.lme4.R
\name{report.lm}
\alias{report.lm}
\alias{report_effectsize.lm}
\alias{report_table.lm}
\alias{report_statistics.lm}
\alias{report_parameters.lm}
\alias{report_intercept.lm}
\alias{report_model.lm}
\alias{report_performance.lm}
\alias{report_info.lm}
\alias{report_text.lm}
\alias{report_random.merMod}
\title{(General) Linear Models Report}
\usage{
\method{report}{lm}(x, include_effectsize = TRUE, effectsize_method = "refit", ...)

\method{report_effectsize}{lm}(x, effectsize_method = "refit", ...)

\method{report_table}{lm}(x, include_effectsize = TRUE, ...)

\method{report_statistics}{lm}(
  x,
  table = NULL,
  include_effectsize = TRUE,
  include_diagnostic = TRUE,
  ...
)

\method{report_parameters}{lm}(
  x,
  table = NULL,
  include_effectsize = TRUE,
  include_intercept = TRUE,
  ...
)

\method{report_intercept}{lm}(x, table = NULL, ...)

\method{report_model}{lm}(x, table = NULL, ...)

\method{report_performance}{lm}(x, table = NULL, ...)

\method{report_info}{lm}(
  x,
  effectsize = NULL,
  include_effectsize = FALSE,
  parameters = NULL,
  ...
)

\method{report_text}{lm}(x, table = NULL, ...)

\method{report_random}{merMod}(x, ...)
}
\arguments{
\item{x}{Object of class \code{lm} or \code{glm}.}

\item{include_effectsize}{If \code{FALSE}, won't include effect-size related indices (standardized coefficients, etc.).}

\item{effectsize_method}{See documentation for \code{\link[effectsize:effectsize]{effectsize::effectsize()}}.}

\item{...}{Arguments passed to or from other methods.}

\item{table}{Provide the output of  \code{report_table()} to avoid its re-computation.}

\item{include_diagnostic}{If \code{FALSE}, won't include diagnostic related indices for Bayesian models (ESS, Rhat).}

\item{include_intercept}{If \code{FALSE}, won't include the intercept.}

\item{effectsize}{Provide the output of \code{report_effectsize()} to avoid its re-computation.}

\item{parameters}{Provide the output of \code{report_parameters()} to avoid its re-computation.}
}
\value{
A list-object of class \code{report}, which contains further list-objects
with a short and long description of the model summary, as well as a short
and long table of parameters and fit indices.
}
\description{
Create a report of a (general) linear model (i.e., a regression fitted using \code{lm()} or \code{glm()}.
}
\examples{
library(report)

# Linear models
model <- lm(Sepal.Length ~ Petal.Length * Species, data = iris)
r <- report(model)
r
summary(r)
as.data.frame(r)
summary(as.data.frame(r))

# Logistic models
model <- glm(vs ~ disp, data = mtcars, family = "binomial")
r <- report(model)
r
summary(r)
as.data.frame(r)
summary(as.data.frame(r))

# Mixed models
if(require("lme4")){
  model <- lme4::lmer(Sepal.Length ~ Petal.Length + (1 | Species), data = iris)
  r <- report(model)
  r
  summary(r)
  as.data.frame(r)
  summary(as.data.frame(r))
}
}
\seealso{
Specific components of reports (especially for stats models):
\itemize{
  \item \code{\link{report_table}}
  \item \code{\link{report_parameters}}
  \item \code{\link{report_statistics}}
  \item \code{\link{report_effectsize}}
  \item \code{\link{report_model}}
  \item \code{\link{report_random}}
  \item \code{\link{report_performance}}
  \item \code{\link{report_info}}
  \item \code{\link{report_text}}
}
Other types of reports:
\itemize{
  \item \code{\link{report_system}}
  \item \code{\link{report_packages}}
  \item \code{\link{report_participants}}
  \item \code{\link{report_sample}}
}
Methods:
\itemize{
  \item \code{\link{as.report}}
}
}
