% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.lm.R
\name{report.lm}
\alias{report.lm}
\title{Linear Models Report}
\usage{
\method{report}{lm}(model, ci = 0.95, standardize = TRUE,
  effsize = "cohen1988", performance_in_table = TRUE,
  performance_metrics = "all", bootstrap = FALSE, iterations = 500,
  ...)
}
\arguments{
\item{model}{Object of class \link{lm}.}

\item{ci}{Confidence Interval (CI) level. Default to 0.95 (95\%).}

\item{standardize}{Standardized coefficients. See \code{\link[parameters:model_parameters.lm]{model_parameters}}.}

\item{effsize}{\href{https://easystats.github.io/report/articles/interpret_metrics.html}{Interpret the standardized parameters} using a set of rules. Can be "cohen1988" (default for linear models), "chen2010" (default for logistic models), "sawilowsky2009", NULL, or a custom set of \link{rules}.}

\item{performance_in_table}{Add performance metrics in table.}

\item{performance_metrics}{See \code{\link[performance:model_performance.lm]{model_performance}}.}

\item{bootstrap}{See \code{\link[parameters:model_parameters.lm]{model_parameters}}.}

\item{iterations}{The number of bootstrap replicates. This only apply in the case of bootsrapped frequentist models.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Create a report of a linear model.
}
\examples{
model <- lm(Sepal.Length ~ Petal.Length * Species, data = iris)
r <- report(model)
to_text(r)
to_fulltext(r)
to_table(r)
to_fulltable(r)
}
