% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.lm.R
\name{report.lm}
\alias{report.lm}
\title{(General) Linear Models Report}
\usage{
\method{report}{lm}(x, ...)
}
\arguments{
\item{x}{Object of class \code{lm} or \code{glm}.}

\item{...}{Arguments passed to or from other methods.}
}
\value{
A list-object of class \code{report}, which contains further list-objects
with a short and long description of the model summary, as well as a short
and long table of parameters and fit indices.
}
\description{
Create a report of a (general) linear model (i.e., a regression fitted using \code{lm()} or \code{glm()}.
}
\examples{
library(report)

model <- lm(Sepal.Length ~ Petal.Length * Species, data = iris)
r <- report(model)
r

# model <- glm(vs ~ disp, data = mtcars, family = "binomial")
# r <- report(model)
}
\seealso{
Aspects of reports:
\itemize{
  \item \code{\link{report_table}}
  \item \code{\link{report_parameters}}
  \item \code{\link{report_statistics}}
  \item \code{\link{report_effectsize}}
  \item \code{\link{report_model}}
  \item \code{\link{report_text}}
}
Methods:
\itemize{
  \item \code{\link{as.report}}
}
Specific reports:
\itemize{
  \item \code{\link{report_system}}
  \item \code{\link{report_packages}}
  \item \code{\link{report_participants}}
  \item \code{\link{report_sample}}
}
}
