% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.glm.R
\name{report.lm}
\alias{report.lm}
\title{(General) Linear Models Report}
\usage{
\method{report}{lm}(
  model,
  effsize = "default",
  ci = 0.95,
  standardize = "refit",
  standardize_robust = FALSE,
  bootstrap = FALSE,
  iterations = 500,
  performance_metrics = "all",
  ...
)
}
\arguments{
\item{model}{Model object.}

\item{effsize}{\href{https://easystats.github.io/report/articles/interpret_metrics.html}{Interpret the standardized parameters} using a set of rules. Default corresponds to "funder2019" for linear models and "chen2010" for logistic models.}

\item{ci}{Confidence Interval (CI) level. Default to 0.95 (95\%).}

\item{standardize}{The method used for standardizing the parameters. Can be \code{"refit"}, \code{"posthoc"}, \code{"smart"}, \code{"basic"} or \code{NULL} (default) for no standardization. See 'Details' in \code{\link[effectsize]{standardize_parameters}}. Note that robust estimation (i.e. \code{robust=TRUE}) of standardized parameters only works when \code{standardize="refit"}.}

\item{bootstrap}{Should estimates be based on bootstrapped model? If \code{TRUE}, then arguments of \link[parameters:model_parameters.stanreg]{Bayesian regressions} apply (see also \code{\link[parameters:parameters_bootstrap]{parameters_bootstrap()}}).}

\item{iterations}{The number of bootstrap replicates. This only apply in the case of bootstrapped frequentist models.}

\item{performance_metrics}{See \code{\link[performance:model_performance.lm]{model_performance}}.}

\item{...}{Arguments passed to or from other methods. For instance, when \code{bootstrap = TRUE}, arguments like \code{ci_method} are passed down to \code{\link[bayestestR]{describe_posterior}}.}
}
\description{
Create a report of a (general) linear model.
}
\examples{
library(report)

model <- lm(Sepal.Length ~ Petal.Length * Species, data = iris)
r <- report(model)
to_text(r)
to_fulltext(r)
to_table(r)
to_fulltable(r)


model <- glm(vs ~ disp, data = mtcars, family = "binomial")
r <- report(model)
to_text(r)
to_fulltext(r)
to_table(r)
to_fulltable(r)
}
