% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.merMod.R
\name{report.lmerMod}
\alias{report.lmerMod}
\title{Mixed Models Report}
\usage{
\method{report}{lmerMod}(model, effsize = "default", ci = 0.95,
  standardize = "refit", standardize_robust = FALSE,
  bootstrap = FALSE, iterations = 500, performance_metrics = "all",
  p_method = "wald", ci_method = "wald", ...)
}
\arguments{
\item{model}{Frequentist regression object.}

\item{effsize}{\href{https://easystats.github.io/report/articles/interpret_metrics.html}{Interpret the standardized parameters} using a set of rules. Default corresponds to "funder2019" for linear models and "chen2010" for logistic models.}

\item{ci}{Confidence Interval (CI) level. Default to 0.95 (95\%).}

\item{standardize}{Add standardized parameters. Can be FALSE or a character indicating the standardization method (see \code{\link{parameters_standardize}}).}

\item{standardize_robust}{Robust standardization. See \code{\link{parameters_standardize}}.}

\item{bootstrap}{Should estimates be based on bootsrapped model? If TRUE, then arguments of \link[=model_parameters.stanreg]{Bayesian regressions} apply.}

\item{iterations}{The number of bootstrap replicates. This only apply in the case of bootsrapped frequentist models.}

\item{performance_metrics}{See \code{\link[performance:model_performance.lm]{model_performance}}.}

\item{p_method}{Method for computing p values. See \link{p_value}.}

\item{ci_method}{Method for computing confidence intervals (CI). See \link{ci}.}

\item{...}{Arguments passed to or from other methods (e.g., to \link[=standardize.lm]{standardize}).}
}
\description{
Create a report of a mixed model (lme4).
}
\examples{
library(lme4)
library(report)

model <- lmer(Sepal.Length ~ Petal.Length + (1 | Species), data = iris)
r <- report(model)
to_text(r)
to_fulltext(r)
to_table(r)
to_fulltable(r)


model <- glmer(vs ~ disp + (1 | am), data = mtcars, family = "binomial")
r <- report(model)
to_text(r)
to_fulltext(r)
to_table(r)
to_fulltable(r)
}
