% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.merMod.R
\name{report.lmerMod}
\alias{report.lmerMod}
\title{Mixed Models Report}
\usage{
\method{report}{lmerMod}(model, effsize = "default", ci = 0.95,
  standardize = "refit", standardize_robust = FALSE,
  bootstrap = FALSE, iterations = 500, performance_metrics = "all",
  p_method = "wald", ci_method = "wald", ...)
}
\arguments{
\item{model}{Model object.}

\item{effsize}{\href{https://easystats.github.io/report/articles/interpret_metrics.html}{Interpret the standardized parameters} using a set of rules. Default corresponds to "funder2019" for linear models and "chen2010" for logistic models.}

\item{ci}{Confidence Interval (CI) level. Default to 0.95 (95\%).}

\item{standardize}{Add standardized parameters. Can be \code{FALSE} or a character indicating the standardization method (see \code{\link[=parameters_standardize]{parameters_standardize()}}), such as \code{"refit"}, \code{"2sd"}, \code{"smart"} or \code{"classic"}. The two formers are based on model refitting using a standardized version of data. It is the most accurate, altough computationnaly heavy (as it must re-fit a second model). The "smart" and "classic" are post-hoc methods, fast, but inaccurate (especially if the model includes interactions).}

\item{standardize_robust}{Robust standardization. See \code{\link{parameters_standardize}}.}

\item{bootstrap}{Should estimates be based on bootsrapped model? If \code{TRUE}, then arguments of \link[=model_parameters.stanreg]{Bayesian regressions} apply (see also \code{\link[=parameters_bootstrap]{parameters_bootstrap()}}).}

\item{iterations}{The number of bootstrap replicates. This only apply in the case of bootsrapped frequentist models.}

\item{performance_metrics}{See \code{\link[performance:model_performance.lm]{model_performance}}.}

\item{p_method}{Method for computing p values. See \code{\link[=p_value]{p_value()}}.}

\item{ci_method}{Method for computing confidence intervals (CI). See \code{\link[=ci]{ci()}}.}

\item{...}{Arguments passed to or from other methods (e.g., to \code{\link[=standardize.lm]{standardize()}}).}
}
\description{
Create a report of a mixed model (lme4).
}
\examples{
library(lme4)
library(report)

model <- lmer(Sepal.Length ~ Petal.Length + (1 | Species), data = iris)
r <- report(model)
to_text(r)
to_fulltext(r)
to_table(r)
to_fulltable(r)


model <- glmer(vs ~ disp + (1 | am), data = mtcars, family = "binomial")
r <- report(model)
to_text(r)
to_fulltext(r)
to_table(r)
to_fulltable(r)
}
