% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_regression.R
\name{report.lmerMod}
\alias{report.lmerMod}
\title{Mixed Models Report}
\usage{
\method{report}{lmerMod}(
  model,
  interpretation = "default",
  ci = 0.95,
  standardize = NULL,
  standardize_robust = FALSE,
  bootstrap = FALSE,
  iterations = 500,
  performance_metrics = "all",
  df_method = "wald",
  ...
)
}
\arguments{
\item{model}{Model object.}

\item{interpretation}{\href{https://easystats.github.io/effectsize/articles/interpret.html}{Interpret the standardized parameters} using a set of rules. Default corresponds to "funder2019" for linear models and "chen2010" for logistic models.}

\item{ci}{Confidence Interval (CI) level. Default to 0.95 (95\%).}

\item{standardize}{The method used for standardizing the parameters. Can be \code{"refit"}, \code{"posthoc"}, \code{"smart"}, \code{"basic"} or \code{NULL} (default) for no standardization. See 'Details' in \code{\link[effectsize]{standardize_parameters}}. Note that robust estimation (i.e. \code{robust=TRUE}) of standardized parameters only works when \code{standardize="refit"}.}

\item{standardize_robust}{Logical, if \code{TRUE}, robust standard errors are calculated (if possible), and confidence intervals and p-values are based on these robust standard errors.}

\item{bootstrap}{Should estimates be based on bootstrapped model? If \code{TRUE}, then arguments of \link[parameters:model_parameters.stanreg]{Bayesian regressions} apply (see also \code{\link[parameters:bootstrap_parameters]{bootstrap_parameters()}}).}

\item{iterations}{The number of bootstrap replicates. This only apply in the case of bootstrapped frequentist models.}

\item{performance_metrics}{See \code{\link[performance:model_performance.lm]{model_performance}}.}

\item{df_method}{Method for computing degrees of freedom for p values, standard errors and confidence intervals (CI). May be \code{"wald"} (default, see \code{\link[parameters]{degrees_of_freedom}}), \code{"ml1"} (see \code{\link[parameters]{dof_ml1}}), \code{"betwithin"} (see \code{\link[parameters]{dof_betwithin}}), \code{"satterthwaite"} (see \code{\link[parameters]{dof_satterthwaite}}) or \code{"kenward"} (see \code{\link[parameters]{dof_kenward}}). Note that when \code{df_method} is not \code{"wald"}, robust standard errors etc. cannot be computed.}

\item{...}{Arguments passed to or from other methods. For instance, when \code{bootstrap = TRUE}, arguments like \code{ci_method} are passed down to \code{\link[bayestestR]{describe_posterior}}.}
}
\description{
Create a report of a mixed model (lme4).
}
\examples{
library(lme4)
library(report)

model <- lmer(Sepal.Length ~ Petal.Length + (1 | Species), data = iris)
r <- report(model)
text_short(r)
text_long(r)
table_short(r)
table_long(r)


model <- glmer(vs ~ disp + (1 | am), data = mtcars, family = "binomial")
r <- report(model)
text_short(r)
text_long(r)
table_short(r)
table_long(r)
}
