% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.numeric.R
\name{report.numeric}
\alias{report.numeric}
\title{Numeric Vector Report}
\usage{
\method{report}{numeric}(x, median = FALSE, dispersion = TRUE,
  range = TRUE, missing_percentage = FALSE, ...)
}
\arguments{
\item{x}{a numeric vector.}

\item{median}{show \link{mean} and \link{sd} (default) or \link{median} and \link{mad}.}

\item{dispersion}{show dispersion (\link{sd} or \link{mad}).}

\item{range}{show range.}

\item{missing_percentage}{show missings by number (default) or percentage}

\item{...}{arguments passed to or from other methods.}
}
\description{
Create a report of a numeric vector.
}
\examples{
x <- rnorm(1000)
report(x)
report(x, median = TRUE, dispersion = TRUE, range = TRUE, missing_percentage = TRUE)
to_fulltext(report(x))
to_table(report(x))
to_fulltable(report(x))
}
\seealso{
report
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
