% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Create Reports}
\usage{
report(x, ...)
}
\arguments{
\item{x}{an R object.}

\item{...}{arguments passed to or from other methods.}
}
\description{
Create textual reports. See the documentation for your object's class:
\itemize{
 \item{\link[=report.numeric]{report.numeric}}
 }
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
