% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Automatic report of R objects}
\usage{
report(model, ...)
}
\arguments{
\item{model}{A statistical model. For data frame or character methods, a data frame or character respectively.}

\item{...}{Arguments passed to or from other methods.}
}
\value{
A list-object of class \code{report}, which contains further list-objects
with a short and long description of the model summary, as well as a short
and long table of parameters and fit indices.
}
\description{
Create reports of different objects. See the documentation for your object's class:
}
\details{
\itemize{
 \item{\link[=report.data.frame]{Dataframes and vectors}}
 \item{\link[=report.htest]{Correlations and t-tests} (\code{htest})}
 \item{\link[=report.aov]{ANOVAs}}
 \item{\link[=report.default]{Regression models} (\code{glm, lm, ...})}
 \item{\link[=report.lmerMod]{Mixed models} \code{(glmer, lmer, glmmTMB, ...)}}
 \item{\link[=report.stanreg]{Bayesian models} \code{(stanreg, brms...)}}
 }



\subsection{Change output type}{
The \code{report()} function generates a report-object that contain in itself
different representations (e.g., text, tables, plots). These different representations
can be accessed via several functions, such as:

\itemize{
\item \strong{\code{text_long()}}: Detailed text.
\item \strong{\code{text_short()}}: Minimal text giving the minimal information.
\item \strong{\code{table_long()}}: Comprehensive table including most available indices.
\item \strong{\code{table_short()}}: Minimal table.
}

Note that for some report objects, some of these representations might be identical.
}
}
\examples{
library(report)

model <- t.test(Sepal.Length ~ Species, data = iris[1:100, ])
r <- report(model)

# Text
r
summary(r)

# Tables
as.data.frame(r)
summary(as.data.frame(r)) # equivalent to as.table(r)

# List
as.list(r)
}
\seealso{
\code{\link{table_short}} or \code{\link{text_short}} to access the
related content of the report-object.
}
