% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Create Reports}
\usage{
report(model, ...)
}
\arguments{
\item{model}{Object.}

\item{...}{Arguments passed to or from other methods.}
}
\value{
A list-object of class \code{report}, which contains further list-objects
with a short and long description of the model summary, as well as a short
and long table of parameters and fit indices.
}
\description{
Create textual reports. See the documentation for your object's class:
\itemize{
 \item{\link[=report.data.frame]{Dataframes and vectors}}
 \item{\link[=report.htest]{Correlations and t-tests} (\code{htest})}
 \item{\link[=report.aov]{ANOVAs}}
 \item{\link[=report.default]{(Generalized) Linear models} (\code{glm, lm, ...})}
 \item{\link[=report.lmerMod]{Mixed models} \code{(glmer, lmer, glmmTMB, ...)}}
 \item{\link[=report.stanreg]{Bayesian models} \code{(stanreg, brms...)}}
 }
}
\examples{
library(report)

model <- t.test(Sepal.Length ~ Species, data = iris[1:100, ])
r <- report(model)

# Text
r
summary(r)

# Tables
as.data.frame(r)
summary(as.data.frame(r)) # equivalent to as.table(r)

# List
as.list(r)
}
\seealso{
\code{\link{table_short}} or \code{\link{text_short}} to access the
related content of the report-object.
}
