% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Create Reports}
\usage{
report(model, ...)
}
\arguments{
\item{model}{Object.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Create textual reports. See the documentation for your object's class:
\itemize{
 \item{\link[=report.data.frame]{Dataframes and vectors}}
 \item{\link[=report.htest]{Correlations and t-tests (htest)}}
 \item{\link[=report.aov]{ANOVAs}}
 \item{\link[=report.glm]{(General) Linear models (glm and lm)}}
 \item{\link[=report.merMod]{Mixed models (glmer and lmer)}}
 \item{\link[=report.stanreg]{Bayesian models (stanreg and brms)}}
 }
}
\examples{
library(report)

model <- t.test(Sepal.Length ~ Species, data = iris[1:100, ])
r <- report(model)

# Text
r
summary(r)

# Tables
as.data.frame(r)
summary(as.data.frame(r)) # equivalent to as.table(r)

# List
as.list(r)
}
