% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Create Reports}
\usage{
report(model, ...)
}
\arguments{
\item{model}{Object.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Create textual reports. See the documentation for your object's class:
\itemize{
 \item{\link[=report.data.frame]{Dataframes and vectors}}
 \item{\link[=report.htest]{Correlations and t-tests (htest)}}
 \item{\link[=report.aov]{ANOVAs}}
 \item{\link[=report.lm]{(General) Linear models (glm and lm)}}
 \item{\link[=report.lmerMod]{Mixed models (glmer and lmer)}}
 \item{\link[=report.stanreg]{Bayesian models (stanreg and brms)}}
 }
}
