% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.stanreg.R
\name{report.stanreg}
\alias{report.stanreg}
\title{Bayesian Models Report}
\usage{
\method{report}{stanreg}(model, ci = 0.95, standardize = TRUE,
  effsize = "cohen1988", performance_in_table = TRUE,
  performance_metrics = "all", estimate = "median", ...)
}
\arguments{
\item{model}{Object of class \link{lm}.}

\item{ci}{Confidence Interval (CI) level. Default to 0.95 (95\%).}

\item{standardize}{Standardized coefficients. See \code{\link[parameters:model_parameters.lm]{model_parameters}}.}

\item{effsize}{Interpret the standardized parameters using a set of rules. Can be "cohen1988" (default), "sawilowsky2009", NULL, or a custom set of \link{rules}.}

\item{performance_in_table}{Add performance metrics on table.}

\item{performance_metrics}{See \code{\link[performance:model_performance.lm]{model_performance}}.}

\item{estimate}{See \code{\link[parameters:model_parameters.lm]{model_parameters}}.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Create a report of a linear model.
}
\examples{
\dontrun{
library(rstanarm)
model <- rstanarm::stan_glm(Sepal.Length ~ Petal.Length * Species, data = iris)
r <- report(model)
to_text(r)
to_fulltext(r)
to_table(r)
to_fulltable(r)
}
}
