% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.stanreg.R
\name{report.stanreg}
\alias{report.stanreg}
\title{Bayesian Models Report}
\usage{
\method{report}{stanreg}(model, ci = 0.9, standardize = FALSE,
  effsize = NULL, performance_in_table = TRUE,
  performance_metrics = c("R2", "R2_adjusted"),
  parameters_estimate = "median", parameters_test = c("pd", "rope"),
  parameters_diagnostic = TRUE, parameters_priors = TRUE,
  rope_range = "default", rope_full = TRUE, ...)
}
\arguments{
\item{model}{Object of class \link{lm}.}

\item{ci}{\href{https://easystats.github.io/bayestestR/articles/1_IndicesDescription.html#hdi---the-credible-interval-ci}{Credible Interval} (CI) level. Default to 0.90 (90\%).}

\item{standardize}{Standardized coefficients. See \code{\link[parameters:model_parameters.lm]{model_parameters}}.}

\item{effsize}{\href{https://easystats.github.io/report/articles/interpret_metrics.html}{Interpret the standardized parameters} using a set of rules. Can be "cohen1988" (default for linear models), "chen2010" (default for logistic models), "sawilowsky2009", NULL, or a custom set of \link{rules}.}

\item{performance_in_table}{Add performance metrics in table.}

\item{performance_metrics}{Can be \code{"all"} or a list of metrics to be computed (some of \code{c("LOO", "R2", "R2_adj")}).}

\item{parameters_estimate}{The \href{https://easystats.github.io/bayestestR/articles/2_IndicesEstimationComparison.html}{point-estimate(s)} to compute. Can be a character or a list with "median", "mean" or "MAP".}

\item{parameters_test}{What \href{https://easystats.github.io/bayestestR/articles/3_IndicesExistenceComparison.html}{indices of effect existence} to compute. Can be a character or a list with "p_direction", "rope" or "p_map".}

\item{parameters_diagnostic}{Include sampling diagnostic metrics (effective sample, Rhat and MCSE). \code{Effective Sample} should be as large as possible, altough for most applications, an effective sample size greater than 1,000 is sufficient for stable estimates (Bürkner, 2017). \code{Rhat} should not be larger than 1.1.}

\item{parameters_priors}{Include priors specifications information. If set to true (current \code{rstanarm}' default), automatically adjusted priors' scale during fitting  will be displayed.}

\item{rope_range}{\href{https://easystats.github.io/bayestestR/articles/1_IndicesDescription.html#rope}{ROPE's} lower and higher limits Should be a list of two values (e.g., \code{c(-0.1, 0.1)}) or \code{"default"}. If \code{"default"}, the bounds are set to \code{x +- 0.1*SD(response)}.}

\item{rope_full}{If TRUE, use the proportion of the entire posterior distribution for the equivalence test. Otherwise, use the proportion of HDI as indicated by the \code{ci} argument.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Create a report of a linear model.
}
\examples{
\dontrun{
library(rstanarm)
model <- rstanarm::stan_glm(Sepal.Length ~ Petal.Length * Species, data = iris)
r <- report(model)
to_text(r)
to_fulltext(r)
to_table(r)
to_fulltable(r)

model <- rstanarm::stan_lmer(Sepal.Length ~ Petal.Length + (1 | Species), data = iris)
report(model)
}
}
