% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-blr.R
\name{report_blr}
\alias{report_blr}
\title{Logistic regression report}
\usage{
report_blr()
}
\description{
Generate detailed logistic regression report.
}
\examples{
model_data <- blorr::bank_marketing
report_blr()

}
