% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_misc.R
\name{report_date}
\alias{report_date}
\alias{report_story}
\title{Miscellaneous reports}
\usage{
report_date(...)

report_story(...)
}
\arguments{
\item{...}{Arguments passed to or from other methods.}
}
\value{
A list-object of class \code{report}, which contains further
  list-objects with a short and long description of the model summary, as
  well as a short and long table of parameters and fit indices.
}
\description{
Other convenient or totally useless reports.
}
\examples{
library(report)

report_date()
summary(report_date())
report_story()
}
\seealso{
Specific components of reports (especially for stats models):
\itemize{
  \item \code{\link{report_table}}
  \item \code{\link{report_parameters}}
  \item \code{\link{report_statistics}}
  \item \code{\link{report_effectsize}}
  \item \code{\link{report_model}}
  \item \code{\link{report_priors}}
  \item \code{\link{report_random}}
  \item \code{\link{report_performance}}
  \item \code{\link{report_info}}
  \item \code{\link{report_text}}
}
Other types of reports:
\itemize{
  \item \code{\link{report_system}}
  \item \code{\link{report_packages}}
  \item \code{\link{report_participants}}
  \item \code{\link{report_sample}}
  \item \code{\link{report_date}}
}
Methods:
\itemize{
  \item \code{\link{as.report}}
}
Template file for supporting new models:
\itemize{
  \item \code{\link{report.default}}
}
}
