% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-summary.R
\name{report_descriptr}
\alias{report_descriptr}
\title{Summary statistics report}
\usage{
report_descriptr()
}
\description{
Generate detailed summary statistics for each variable/column in a data set.
}
\examples{
library(descriptr)
report_descriptr()

}
