% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_effectsize.R
\name{report_effectsize}
\alias{report_effectsize}
\title{Report the effect size(s) of a model or a test}
\usage{
report_effectsize(x, ...)
}
\arguments{
\item{x}{The R object that you want to report (see list of of supported objects above).}

\item{...}{Arguments passed to or from other methods.}
}
\value{
An object of class \code{report_effectsize}.
}
\description{
Computes, interpret and formats the effect sizes of a variety of models and statistical tests (see list of supported objects in \code{\link{report}}).
}
\examples{
library(report)

# h-tests
report_effectsize(t.test(iris$Sepal.Width, iris$Sepal.Length))
report_effectsize(cor.test(iris$Sepal.Width, iris$Sepal.Length))
}
