% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_info.R
\name{report_info}
\alias{report_info}
\title{Report additional information}
\usage{
report_info(x, table = NULL, ...)
}
\arguments{
\item{x}{The R object that you want to report (see list of of supported objects above).}

\item{table}{A table obtained via \code{report_table()}. If not provided, will run it.}

\item{...}{Arguments passed to or from other methods.}
}
\value{
A \code{character} string.
}
\description{
Reports additional information relevant to the report (see list of supported objects in \code{\link{report}}).
}
\examples{
library(report)

# h-tests
report_info(t.test(iris$Sepal.Width, iris$Sepal.Length))
report_info(cor.test(iris$Sepal.Width, iris$Sepal.Length))
}
