% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_msd.R
\name{report_msd}
\alias{report_msd}
\title{Generate an in-text report of means and standard deviations}
\usage{
report_msd(variable = NULL, format = "rmarkdown", inline = T)
}
\arguments{
\item{variable}{variable to report descriptives for}

\item{format}{How should the output be formatted? Available options are:
\code{"plain"}, \code{"latex"}, \code{"rmarkdown"}}

\item{inline}{Should non-plain results be formatted for inline (\code{TRUE})
or for copy-pasting (\code{FALSE})?
(default is \code{TRUE})}
}
\value{
Mean and SD reported to two decimal places
}
\description{
Generate an in-text report of means and standard deviations
}
\examples{
age <- rnorm(110, 21, 3.4)
report_msd(age)
}
