% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-ols.R
\name{report_ols}
\alias{report_ols}
\title{Linear regression report}
\usage{
report_ols(use_yaml = FALSE)
}
\arguments{
\item{use_yaml}{Use YAML file to specify report details.}
}
\description{
Generate detailed linear regression report.
}
\examples{
\dontrun{
model_data <- descriptr::mtcarz

# interactive
report_ols()

# use yaml file
report_ols(use_yaml = TRUE)
}

}
