% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_p.R
\name{report_p}
\alias{report_p}
\title{Automate the reporting of p-values}
\usage{
report_p(pvalue = NULL, markdown = T)
}
\arguments{
\item{pvalue}{statistical object to report}

\item{markdown}{should the output be formatted for Markdown (e.g., *p* ) or plain text?}
}
\value{
p-value rounded to three decimal places if greater than .001. If less than,
.001 defaults to p < .001.
}
\description{
Automate the reporting of p-values
}
\examples{
report_p(0.001567)
}
