% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_parameters.R
\name{report_parameters}
\alias{report_parameters}
\title{Report the parameters of a model}
\usage{
report_parameters(x, table = NULL, prefix = "  - ", ...)
}
\arguments{
\item{x}{The R object that you want to report (see list of of supported objects above).}

\item{table}{A table obtained via \code{report_table()}. If not provided, will run it.}

\item{prefix}{The prefix to be displayed in front of each parameter.}

\item{...}{Arguments passed to or from other methods.}
}
\value{
A \code{vector}.
}
\description{
Creates a list containing a description of the parameters of R objects (see list of supported objects in \code{\link{report}}).
}
\examples{
library(report)

# Miscellaneous
r <- report_parameters(sessionInfo())
r
summary(r)
}
