% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_participants.R
\name{report_participants}
\alias{report_participants}
\title{Reporting the participant data}
\usage{
report_participants(data, age = "Age", sex = "Sex",
  education = "Education", participants = NULL, group = NULL,
  spell_n = FALSE, ...)
}
\arguments{
\item{data}{A dataframe.}

\item{age}{The name of the column containing the age.}

\item{sex}{The name of the column containing the sex. Note that classes should be some of c("Male", "M", "Female", "F").}

\item{participants}{The name of the participants' identifier column (for instance in the case of repeated measures).}

\item{group}{A character vector indicating the name(s) of the column(s) used for stratified description.}

\item{spell_n}{Fully spell the sample size ("Three participants" instead of "3 participants").}

\item{...}{Arguments passed to or from other methods.}
}
\description{
A helper function to help you format the participants data (age, sex, ...) in the participants section.
}
\examples{
library(report)
data <- data.frame("Age" = c(22, 23, 54, 21, 8, 42),
                   "Sex" = c("F", "F", "M", "M", "M", "F"))

report_participants(data, age = "Age", sex = "Sex")

# Years of education (relative to high school graduation)
data$Education <- c(0, 8, -3, -5, 3, 5)
report_participants(data, age = "Age", sex = "Sex", education = "Education")

# Education as factor
data$Education2 <- c("Bachelor", "PhD", "Highschool",
                     "Highschool", "Bachelor", "Bachelor")
report_participants(data, age = "Age", sex = "Sex", education = "Education2")


# Repeated measures data
data <- data.frame("Age" = c(22, 22, 54, 54, 8, 8),
                   "Sex" = c("F", "F", "M", "M", "F", "F"),
                   "Participant" = c("S1", "S1", "s2", "s2", "s3", "s3"))

report_participants(data, age = "Age", sex = "Sex", participants = "Participant")

# Grouped data
data <- data.frame("Age" = c(22, 22, 54, 54, 8, 8, 42, 42),
                   "Sex" = c("F", "F", "M", "M", "F", "F", "M", "M"),
                   "Participant" = c("S1", "S1", "s2", "s2", "s3", "s3", "s4", "s4"),
                   "Condition" = c("A", "A", "A", "A", "B", "B", "B", "B"))

report_participants(data, age = "Age",
                          sex = "Sex",
                          participants = "Participant",
                          group = "Condition")

# Spell sample size
paste(report_participants(data, participants = "Participant", spell_n = TRUE),
      "were recruited in the study by means of torture and coercion.")

}
