% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_performance.R
\name{report_performance}
\alias{report_performance}
\title{Report the model's performance}
\usage{
report_performance(x, table = NULL, ...)
}
\arguments{
\item{x}{The R object that you want to report (see list of of supported objects above).}

\item{table}{A table obtained via \code{report_table()}. If not provided, will run it.}

\item{...}{Arguments passed to or from other methods.}
}
\value{
A \code{character} string.
}
\description{
Reports the type of different R objects (see list of supported objects in \code{\link{report}}).
}
\examples{
library(report)

# GLMs
report_performance(lm(Sepal.Length ~ Petal.Length * Species, data = iris))
}
