% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_performance.R
\name{report_performance}
\alias{report_performance}
\title{Report the model's performance}
\usage{
report_performance(x, table = NULL, ...)
}
\arguments{
\item{x}{The R object that you want to report (see list of of supported objects above).}

\item{table}{A table obtained via \code{report_table()}. If not provided, will run it.}

\item{...}{Arguments passed to or from other methods.}
}
\value{
A \code{character} string.
}
\description{
Reports the type of different R objects (see list of supported objects in \code{\link{report}}).
}
\examples{
library(report)

# GLMs
report_performance(lm(Sepal.Length ~ Petal.Length * Species, data = iris))
report_performance(glm(vs ~ disp, data = mtcars, family = "binomial"))

# Mixed models
if (require("lme4")) {
  model <- lme4::lmer(Sepal.Length ~ Petal.Length + (1 | Species), data = iris)
  report_performance(model)
}

# Bayesian models
if (require("rstanarm")) {
  model <- stan_glm(Sepal.Length ~ Species, data = iris, refresh = 0, iter = 600)
  report_performance(model)
}

# Structural Equation Models (SEM)
if (require("lavaan")) {
  structure <- " ind60 =~ x1 + x2 + x3
                 dem60 =~ y1 + y2 + y3
                 dem60 ~ ind60 "
  model <- lavaan::sem(structure, data = PoliticalDemocracy)
  report_performance(model)
}
}
