% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-rfm.R
\name{report_rfm}
\alias{report_rfm}
\title{RFM report}
\usage{
report_rfm()
}
\description{
Generate recency, frequency and monetary value analysis report.
}
\examples{
rfm_data <- rfm::rfm_data_orders
report_rfm()

}
