% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_s.R
\name{report_s}
\alias{report_s}
\title{Report S- and p-values in easy language.}
\usage{
report_s(s = NULL, p = NULL)
}
\arguments{
\item{s}{An S-value. Either \code{s} or \code{p} must be provided.}

\item{p}{A p-value. Either \code{s} or \code{p} must be provided.}
}
\description{
Reports interpretation of S- and p-values in easy language.
}
\examples{
report_s(s = 1.5)
report_s(p = 0.05)
}
