% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_table.R
\name{report_table}
\alias{report_table}
\title{Report a descriptive table}
\usage{
report_table(x, ...)
}
\arguments{
\item{x}{The R object that you want to report (see list of of supported objects above).}

\item{...}{Arguments passed to or from other methods.}
}
\value{
A \code{data.frame}.
}
\description{
Creates tables to describe different objects (see list of supported objects in \code{\link{report}}).
}
\examples{
library(report)

# Miscellaneous
r <- report_table(sessionInfo())
r
summary(r)

# Data
report_table(iris$Sepal.Length)

# Tests
# report_table(t.test(mpg ~ am, data = mtcars))
}
