% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_text.R
\name{report_text}
\alias{report_text}
\title{Report a textual description of an object}
\usage{
report_text(x, table = NULL, ...)
}
\arguments{
\item{x}{The R object that you want to report (see list of of supported objects above).}

\item{table}{A table obtained via \code{report_table()}. If not provided, will run it.}

\item{...}{Arguments passed to or from other methods.}
}
\value{
A \code{character} string.
}
\description{
Creates text containing a description of the parameters of R objects (see list of supported objects in \code{\link{report}}).
}
\examples{
library(report)

# Miscellaneous
r <- report_text(sessionInfo())
r
summary(r)
}
