% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cite_packages.R
\name{show_packages}
\alias{show_packages}
\alias{cite_packages}
\title{Cite Loaded Packages}
\usage{
show_packages(session)

cite_packages(session)
}
\arguments{
\item{session}{A \link[=sessionInfo]{sessionInfo} object.}
}
\description{
Citation table of loaded packages (\link{show_packages} includes version and name, and \link{cite_packages} includes only the citation).
}
\examples{
show_packages(sessionInfo())
cite_packages(sessionInfo())
}
