% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{text_long}
\alias{text_long}
\alias{text_short}
\alias{table_long}
\alias{table_short}
\title{Modify report output.}
\usage{
text_long(r, ...)

text_short(r, ...)

table_long(r, ...)

table_short(r, ...)
}
\arguments{
\item{r}{Object of class \link{report}.}

\item{...}{Arguments passed to or from other methods.}
}
\value{
Character string or dataframe depending on the method.
}
\description{
Functions to access different representations of a report-object.
See the details section below.
}
\details{
\itemize{
 \item{\link[=report.data.frame]{Dataframes and vectors}}
 \item{\link[=report.htest]{Correlations and t-tests} (\code{htest})}
 \item{\link[=report.aov]{ANOVAs}}
 \item{\link[=report.default]{Regression models} (\code{glm, lm, ...})}
 \item{\link[=report.lmerMod]{Mixed models} \code{(glmer, lmer, glmmTMB, ...)}}
 \item{\link[=report.stanreg]{Bayesian models} \code{(stanreg, brms...)}}
 }



\subsection{Change output type}{
The \code{report()} function generates a report-object that contain in itself
different representations (e.g., text, tables, plots). These different representations
can be accessed via several functions, such as:

\itemize{
\item \strong{\code{text_long()}}: Detailed text.
\item \strong{\code{text_short()}}: Minimal text giving the minimal information.
\item \strong{\code{table_long()}}: Comprehensive table including most available indices.
\item \strong{\code{table_short()}}: Minimal table.
}

Note that for some report objects, some of these representations might be identical.
}
}
