% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_performance.R
\name{text_performance}
\alias{text_performance}
\title{Performance textual reporting}
\usage{
text_performance(model, performance, ...)
}
\arguments{
\item{model}{Object.}

\item{performance}{Performance table.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Convert performance table to text.
}
\seealso{
report
}
