% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{to_fulltable}
\alias{to_fulltable}
\title{Report table display}
\usage{
to_fulltable(x, full = TRUE, digits = NULL, ...)
}
\arguments{
\item{x}{An object of class \link{report}}

\item{full}{Show the full report.}

\item{digits}{Format the digits.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Report table display
}
