% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{to_table}
\alias{to_table}
\title{Report table display}
\usage{
to_table(object, full = FALSE, digits = NULL, ...)
}
\arguments{
\item{object}{Object of class \link{report}.}

\item{full}{Show the full report.}

\item{digits}{Number of digits.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Report table display
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
