% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{to_text}
\alias{to_text}
\alias{to_table}
\alias{to_values}
\title{Report printing}
\usage{
to_text(x, ...)

to_table(x, ...)

to_values(x, ...)
}
\arguments{
\item{x}{Object of class \link{report}.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Report printing
}
